Virtual reality moral dilemmas.

Virtual reality moral dilemmas.

VR_dilemma

Format

A data frame with 68 rows and 4 variables

id. Dummy identity number for each participant.

order. The order in which the participants completed the two sessions: "text_first" (0) or "text_second" (1).

modality. Describes how the moral dilemmas were presented to the participants: either in text format ("text") or in Virtual Reality ("vr").

score. Proportion of "utilitarian" decisions. In other words, of the 4 decisions, how many affirmative were responses. Range: 0 (all utilitarian) - 1 (none utilitarian).

Source

https://psyarxiv.com/ry3ap/

Details

Dataset from a study where participants completed identical moral dilemmas in two different sessions held on separate days: in one session, they read text description of the scenario, while in another session they completed the same scenarios in Virtual Reality (videos: https://www.youtube.com/watch?v=ebdU3HhhYs8). The study investigated if there was a discrepancy between how people judged the same scenarios while reading them in text versus experiencing them in virtual reality.

Examples

dim(VR_dilemma)

#> [1] 68  4

head(VR_dilemma)

#> # A tibble: 6 x 4
#>      id order modality score
#>   <dbl> <chr> <chr>    <dbl>
#> 1     1 0     text      0.25
#> 2     6 0     text      1   
#> 3     7 0     text      1   
#> 4     8 0     text      1   
#> 5     9 0     text      0.5 
#> 6    10 0     text      0.75

dplyr::glimpse(VR_dilemma)

#> Rows: 68
#> Columns: 4
#> $ id       <dbl> 1, 6, 7, 8, 9, 10, 12, 14, 15, 17, 20, 22, 28, 31, 34, 35, 36~
#> $ order    <chr> "0", "0", "0", "0", "0", "0", "0", "0", "0", "0", "0", "0", "~
#> $ modality <chr> "text", "text", "text", "text", "text", "text", "text", "text~
#> $ score    <dbl> 0.25, 1.00, 1.00, 1.00, 0.50, 0.75, 1.00, 1.00, 0.75, 0.00, 0~