Making aesthetic modifications to the plot

Making aesthetic modifications to the plot

aesthetic_addon(
  plot,
  x,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ggtheme = ggplot2::theme_bw(),
  ggstatsplot.layer = TRUE,
  package = "RColorBrewer",
  palette = "Dark2",
  ggplot.component = NULL,
  ...
)

Arguments

plot
Plot to be aesthetically modified.

x
A numeric vector for x axis.

xlab
Labels for x and y axis variables. If NULL (default), variable names for x and y will be used.

ylab
Labels for x and y axis variables. If NULL (default), variable names for x and y will be used.

title
The text for the plot title.

subtitle
The text for the plot subtitle. Will work only if results.subtitle = FALSE.

caption
The text for the plot caption.

ggtheme
A function, ggplot2 theme name. Default value is ggplot2::theme_bw(). Any of the ggplot2 themes, or themes from extension packages are allowed (e.g., ggthemes::theme_fivethirtyeight(), hrbrthemes::theme_ipsum_ps(), etc.).

ggstatsplot.layer
Logical that decides whether theme_ggstatsplot theme elements are to be displayed along with the selected ggtheme (Default: TRUE). theme_ggstatsplot is an opinionated theme layer that override some aspects of the selected ggtheme.

package
Name of package from which the palette is desired as string or symbol.

palette
Name of palette as string or symbol.

ggplot.component
A ggplot component to be added to the plot prepared by ggstatsplot. This argument is primarily helpful for grouped_ variant of the current function. Default is NULL. The argument should be entered as a function.

...
Additional arguments.