Convenience function to extract bayes factors and details about posterior (mean, median, HDI, etc.) from BayesFactor model object.

Convenience function to extract bayes factors and details about posterior (mean, median, HDI, etc.) from BayesFactor model object.

bf_extractor(bf.object, posterior = FALSE, iterations = 1000,
  cred.int = 0.95)

Arguments

bf.object
An object from BayesFactor package test results.

posterior
If TRUE, the dataframe will contain mean, median, standard deviation, and standard error for the posterior.

iterations
The number of iterations to sample for computing posterior (default: 1000).

cred.int
A scalar between 0 and 1, indicating the mass within the credible interval that is to be estimated (default: 0.95).

Examples

# getting only bayes factors
ggstatsplot::bf_extractor(BayesFactor::anovaBF(Sepal.Length ~ Species,
  data = iris,
  progress = FALSE
))

#> # A tibble: 1 x 7
#>      bf10 log_e_bf10 log_10_bf10     bf01 log_e_bf01 log_10_bf01        error
#>     <dbl>      <dbl>       <dbl>    <dbl>      <dbl>       <dbl>        <dbl>
#> 1 1.43e28       64.8        28.2 7.02e-29      -64.8       -28.2 0.0000000651

# show all columns in a tibble
options(tibble.width = Inf)

# getting bayes factors and posteriors
ggstatsplot::bf_extractor(
  BayesFactor::correlationBF(
    x = iris$Sepal.Length,
    y = iris$Sepal.Width
  ),
  posterior = TRUE,
  iterations = 1000,
  cred.int = 0.95
)

#> # A tibble: 1 x 14
#>    bf10 log_e_bf10 log_10_bf10  bf01 log_e_bf01 log_10_bf01 error posterior.mean
#>   <dbl>      <dbl>       <dbl> <dbl>      <dbl>       <dbl> <dbl>          <dbl>
#> 1 0.509     -0.675      -0.293  1.96      0.675       0.293     0          -0.11
#>   posterior.median posterior.sd posterior.std.error HDI.low HDI.high cred.int
#>              <dbl>        <dbl>               <dbl>   <dbl>    <dbl>    <dbl>
#> 1            -0.12        0.081             0.00181  -0.268   0.0468     0.95