Bayes factor message for random-effects meta-analysis

Bayes factor message for random-effects meta-analysis

bf_meta_message(
  data,
  d = prior("norm", c(mean = 0, sd = 0.3)),
  tau = prior("invgamma", c(shape = 1, scale = 0.15)),
  k = 2,
  caption = NULL,
  messages = TRUE,
  ...
)

Arguments

data
data frame containing the variables for effect size y, standard error SE, labels, and moderators per study.

d
prior distribution on the average effect size d. The prior probability density function is defined via prior.

tau
prior distribution on the between-study heterogeneity tau (i.e., the standard deviation of the study effect sizes dstudy in a random-effects meta-analysis. A (nonnegative) prior probability density function is defined via prior.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

caption
The text for the plot caption.

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

...
Arguments passed on to metaBMA::meta_random

y

effect size per study. Can be provided as (1) a numeric vector, (2) the quoted or unquoted name of the variable in data, or (3) a formula to include discrete or continuous moderator variables.

SE

standard error of effect size for each study. Can be a numeric vector or the quoted or unquoted name of the variable in data

labels

optional: character values with study labels. Can be a character vector or the quoted or unquoted name of the variable in data

rscale_contin

scale parameter of the JZS prior for the continuous covariates.

rscale_discrete

scale parameter of the JZS prior for discrete moderators.

centering

whether continuous moderators are centered.

logml

how to estimate the log-marginal likelihood: either by numerical integration ("integrate") or by bridge sampling using MCMC/Stan samples ("stan"). To obtain high precision with logml="stan", many MCMC samples are required (e.g., logml_iter=10000, warmup=1000).

summarize

how to estimate parameter summaries (mean, median, SD, etc.): Either by numerical integration (summarize = "integrate") or based on MCMC/Stan samples (summarize = "stan").

ci

probability for the credibility/highest-density intervals.

rel.tol

relative tolerance used for numerical integration using integrate. Use rel.tol=.Machine$double.eps for maximal precision (however, this might be slow).

logml_iter

number of iterations (per chain) from the posterior distribution of d and tau. The samples are used for computing the marginal likelihood of the random-effects model with bridge sampling (if logml="stan") and for obtaining parameter estimates (if summarize="stan"). Note that the argument iter=2000 controls the number of iterations for estimation of the random-effect parameters per study in random-effects meta-analysis.

silent_stan

whether to suppress the Stan progress bar.

Examples

# \donttest{
# setup
set.seed(123)
library(metaBMA)

# creating a dataframe
(df <-
  structure(
    .Data = list(
      study = c("1", "2", "3", "4", "5"),
      estimate = c(
        0.382047603321706,
        0.780783111514665,
        0.425607573765058,
        0.558365541235078,
        0.956473848429961
      ),
      std.error = c(
        0.0465576338644502,
        0.0330218199731529,
        0.0362834986178494,
        0.0480571500648261,
        0.062215818388157
      )
    ),
    row.names = c(NA, -5L),
    class = c("tbl_df", "tbl", "data.frame")
  ))

#> # A tibble: 5 x 3
#>   study estimate std.error
#>   <chr>    <dbl>     <dbl>
#> 1 1        0.382    0.0466
#> 2 2        0.781    0.0330
#> 3 3        0.426    0.0363
#> 4 4        0.558    0.0481
#> 5 5        0.956    0.0622

# getting Bayes factor in favor of null hypothesis
ggstatsplot::bf_meta_message(
  data = df,
  k = 3,
  iter = 1500,
  messages = TRUE
)

#> ### Bayesian Random-Effects Meta-Analysis ### 
#>    Prior on d:      'norm' (mean=0, sd=0.3) with support on the interval [-Inf,Inf]. 
#>    Prior on tau:    'invgamma' (shape=1, scale=0.15) with support on the interval [0,Inf]. 
#> 
#> # Bayes factors:
#>            (denominator)
#> (numerator) random_H0 random_H1
#>   random_H0       1.0    0.0354
#>   random_H1      28.2    1.0000
#> 
#> # Posterior summary statistics of random-effects model:
#>      mean    sd  2.5%   50% 97.5% hpd95_lower hpd95_upper n_eff  Rhat
#> d   0.516 0.142 0.174 0.537 0.737       0.216       0.760 773.8 1.003
#> tau 0.292 0.137 0.137 0.256 0.687       0.111       0.557 672.5 1.003

#> atop(displaystyle(NULL), expr = paste("In favor of null: ", "log"["e"], 
#>     "(BF"["01"], ") = ", "-3.341", ", ", italic("d")["mean"]^"posterior", 
#>     " = ", "0.516", ", CI"["95%"], " [", "0.216", ", ", "0.760", 
#>     "]"))

# }