Bayesian one-way analysis of variance.

Bayesian one-way analysis of variance.

bf_oneway_anova(data, x, y, bf.prior = 0.707, caption = NULL,
  output = "null", paired = FALSE, k = 2, ...)

Arguments

data
a data frame containing data for all factors in the formula

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

caption
The text for the plot caption.

output
Can either be "null" (or "caption" or "H0", which will contain text for evidence in favor of the null hypothesis or H0) or "alternative" (or "title" or "H1") or "results", which will return a dataframe with results all the details).

paired
a logical indicating whether you want a paired t-test.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

...
Additional arguments.

See also

Examples

# to get caption (default)
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8
)

#> atop(displaystyle(NULL), expr = paste("In favor of null: ", "log"["e"], 
#>     "(BF"["01"], ") = ", "-65.18", ", ", italic("r")["Cauchy"], 
#>     " = ", "0.80"))

# to get results dataframe
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8,
  output = "results"
)

#> # A tibble: 1 x 8
#>      bf10 log_e_bf10 log_10_bf10     bf01 log_e_bf01 log_10_bf01   error
#>     <dbl>      <dbl>       <dbl>    <dbl>      <dbl>       <dbl>   <dbl>
#> 1 2.03e28       65.2        28.3 4.93e-29      -65.2       -28.3 4.78e-8
#> # ... with 1 more variable: bf.prior <dbl>