Wide-format version of the "Bugs" dataset.

Wide-format version of the "Bugs" dataset.

bugs_wide

Format

A data frame with 93 rows and 6 variables

subject. Dummy identity number for each participant.

gender. Participant's gender (Female, Male).

region. Region of the world the participant was from.

education. Level of education.

ldlf,ldhf,hdlf,hdhf. The desire to kill an arthropod was indicated on a scale from 0 to 10 in each condition of the experiment (LDLF: low freighteningness and low disgustingness; LFHD: low freighteningness and high disgustingness; HFHD: high freighteningness and low disgustingness; HFHD: high freighteningness and high disgustingness).

Source

https://www.sciencedirect.com/science/article/pii/S0747563213000277

Details

This data set, "Bugs", provides the extent to which men and women want to kill arthropods that vary in freighteningness (low, high) and disgustingness (low, high). Each participant rates their attitudes towards all anthropods. Subset of the data reported by Ryan et al. (2013).

Examples

dim(bugs_wide)
#> [1] 93  8
head(bugs_wide)
#> # A tibble: 6 × 8
#>   subject gender region        education  ldlf  ldhf  hdlf  hdhf
#>     <int> <fct>  <fct>         <fct>     <dbl> <dbl> <dbl> <dbl>
#> 1       1 Female North America some          6   6     9    10  
#> 2       2 Female North America advance      10  NA    10    10  
#> 3       3 Female Europe        college       5  10    10    10  
#> 4       4 Female North America college       6   9     6     9  
#> 5       5 Female North America some          3   6.5   5.5   8.5
#> 6       6 Female Europe        some          2   0.5   7.5   3  
dplyr::glimpse(bugs_wide)
#> Rows: 93
#> Columns: 8
#> $ subject   <int> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 1…
#> $ gender    <fct> Female, Female, Female, Female, Female, Female, Female, Fema…
#> $ region    <fct> North America, North America, Europe, North America, North A…
#> $ education <fct> some, advance, college, college, some, some, some, high, hig…
#> $ ldlf      <dbl> 6.0, 10.0, 5.0, 6.0, 3.0, 2.0, 10.0, 10.0, 9.5, 8.5, 0.0, 9.…
#> $ ldhf      <dbl> 6.0, NA, 10.0, 9.0, 6.5, 0.5, 10.0, 10.0, 9.5, 10.0, 2.5, 7.…
#> $ hdlf      <dbl> 9.0, 10.0, 10.0, 6.0, 5.5, 7.5, 10.0, 9.0, 6.0, 7.0, 0.0, 8.…
#> $ hdhf      <dbl> 10.0, 10.0, 10.0, 9.0, 8.5, 3.0, 10.0, 10.0, 10.0, NA, 0.0, …