Summary dataframe for categorical variables.

Creating a dataframe with an added column corresponding to summary for categorical variables.

cat_label_df(data, label.col.name = "slice.label",
  label.content = "percentage", label.separator = c("\n", " "),
  perc.k = 1)

Arguments

data
A dataframe containing summaries for categorical variables. Should contain columns named either "perc" or "counts" or both.

label.col.name
Character that decides the column name containing summary label. This can either be "slice.label" (default) or "data.label".

label.content
Character decides what information needs to be displayed on the label in each pie or bar slice. Possible options are "percentage" (default), "counts", "both".

label.separator
If "both" counts and proportion information is to be displayed in a label, this argument decides whether these two pieces of information are going to be on the same line (" ") or on separate lines ("\n").

perc.k
Numeric that decides number of decimal places for percentage labels (Default: 0).

Examples

# dataframe with label column
ggstatsplot:::cat_label_df(
  data = ggstatsplot:::cat_counter(mtcars, am, cyl),
  label.col.name = "slice.label",
  label.content = "both",
  perc.k = 1
)

#> # A tibble: 6 x 5
#>     cyl    am counts  perc slice.label     
#>   <dbl> <dbl>  <int> <dbl> <chr>           
#> 1     4     1      8  72.7 "n = 8\n(72.7%)"
#> 2     6     1      3  42.9 n = 3 (42.9%)   
#> 3     8     1      2  14.3 "n = 2\n(14.3%)"
#> 4     4     0      3  27.3 n = 3 (27.3%)   
#> 5     6     0      4  57.1 "n = 4\n(57.1%)"
#> 6     8     0     12  85.7 n = 12 (85.7%)