Cramer's V and confidence intervals for a chi squared goodness of fit test using non central parameters.

A customized version of DescTools::CramerV to allow for probabilities to be specified. Called from helpers_ggpiestats_subtitle after the appropriate x has been made into a table.

cramer_v_ci(x, conf.level = 0.95, method = c("ncchisq", "ncchisqadj"),
  p = rep(1/length(x), length(x)))

Arguments

x
table of counts.

conf.level
confidence level of the interval around Cramer's V.

method
string defining the method to calculate confidence intervals for Cramer's V. Either "ncchisq" (using noncentral chisquare), or the adjusted version "ncchisqadj"`` with degrees of freedom added to both the numerator and denominator. Default is "ncchisq"`.

p
(passed to chisq.test) a vector of probabilities of the same length of x. p is rescaled (if necessary and it is possible) to sum to

An error is given if any entry of p is negative. Default is equal counts per category (factor level).

Examples

# to get reproducible results from bootstrapping
set.seed(123)
library(ggstatsplot)

# simple function call with the defaults
x <- table(mtcars$am)
ggstatsplot:::cramer_v_ci(x)

#>  Cramer V    lwr.ci    upr.ci 
#> 0.1875000 0.0000000 0.5338788 

# raise confidence intervals specify proportions
ggstatsplot:::cramer_v_ci(x, conf.level = 0.999, p = c(0.6, 0.4))

#>   Cramer V     lwr.ci     upr.ci 
#> 0.01275776 0.00000000 0.54508636 

# notice order matters and it follows the levels of `x`
# in this case am = 0 is before am = 1
ggstatsplot:::cramer_v_ci(x, conf.level = 0.90, p = c(4, 6))

#>  Cramer V    lwr.ci    upr.ci 
#> 0.3954905 0.1003035 0.6862628 

# function body