Convert easystats package outputs to tidymodels conventions.

Both broom package from tidymodels universe and parameters package from easystats universe can provide model summaries for a large number of model objects. This is a convenience function that converts naming conventions adopted in easystats to the ones adopted in the broom package.

easystats_to_tidy_names(x)

Examples

# example model object
mod <- stats::lm(formula = wt ~ am * cyl, data = mtcars)

# `tidy`-fied output
ggstatsplot:::easystats_to_tidy_names(parameters::model_parameters(mod))

#> # A tibble: 4 x 8
#>   term        estimate std.error conf.low conf.high statistic df.error  p.value
#>   <chr>          <dbl>     <dbl>    <dbl>     <dbl>     <dbl>    <int>    <dbl>
#> 1 (Intercept)   1.66      0.587     0.455     2.86      2.82        28 0.00864 
#> 2 am           -0.956     0.793    -2.58      0.668    -1.21        28 0.238   
#> 3 cyl           0.304     0.0826    0.135     0.473     3.68        28 0.000989
#> 4 am:cyl        0.0328    0.130    -0.234     0.300     0.252       28 0.803