Extracting dataframes with statistical details from ggstatsplot

Extracting dataframes with statistical details from ggstatsplot

extract_stats(p, ...)

Arguments

p
A plot from ggstatsplot package

...
Ignored

Details

This is a convenience function to extract dataframes with statistical details that are used to create expressions displayed in ggstatsplot plots as subtitle and/or as caption. Note that all of this analysis is carried out by the statsExpressions package.

For more details about underlying tests and effect size estimates, see the following vignette: https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Note

Note that if you want to use this function with ggscatterstats, you will have to set marginal = FALSE to return an object of ggplot type.

Examples

# \donttest{
set.seed(123)
library(ggstatsplot)
p <- ggbetweenstats(mtcars, am, mpg)
extract_stats(p)
#> $subtitle_data
#> # A tibble: 1 x 16
#>   term  group mean.group1 mean.group2 statistic df.error p.value method estimate
#>   <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl> <chr>     <dbl>
#> 1 mpg   am           17.1        24.4     -3.77     18.3 0.00137 Welch…    -1.35
#> # … with 7 more variables: conf.level <dbl>, conf.low <dbl>, conf.high <dbl>,
#> #   effectsize <chr>, conf.method <chr>, conf.distribution <chr>,
#> #   expression <list>
#> 
#> $caption_data
#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference     6.44       0.95     2.68    10.0   0.999               0
#> 2 Cohens_d      -1.30       0.95    -2.11    -0.519 0.999               0
#> # … with 6 more variables: prior.distribution <chr>, prior.location <dbl>,
#> #   prior.scale <dbl>, bf10 <dbl>, method <chr>, expression <list>
#> 
# }