Split dataframe into a list by grouping variable.

This function splits the dataframe into a list, with the length of the list equal to the factor levels of the grouping variable. Each element of the list will be a tibble.

grouped_list(data, grouping.var = NULL)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

grouping.var
A single grouping variable.

Examples

# \donttest{
ggstatsplot:::grouped_list(ggplot2::msleep, grouping.var = vore)
#> $carni
#> # A tibble: 19 × 11
#>    name      genus  vore  order  conservation sleep_total sleep_rem sleep_cycle awake brainwt
#>    <chr>     <chr>  <chr> <chr>  <chr>              <dbl>     <dbl>       <dbl> <dbl>   <dbl>
#>  1 Cheetah   Acino… carni Carni… lc                  12.1      NA        NA      11.9 NA     
#>  2 Northern… Callo… carni Carni… vu                   8.7       1.4       0.383  15.3 NA     
#>  3 Dog       Canis  carni Carni… domesticated        10.1       2.9       0.333  13.9  0.07  
#>  4 Long-nos… Dasyp… carni Cingu… lc                  17.4       3.1       0.383   6.6  0.0108
#>  5 Domestic… Felis  carni Carni… domesticated        12.5       3.2       0.417  11.5  0.0256
#>  6 Pilot wh… Globi… carni Cetac… cd                   2.7       0.1      NA      21.4 NA     
#>  7 Gray seal Halio… carni Carni… lc                   6.2       1.5      NA      17.8  0.325 
#>  8 Thick-ta… Lutre… carni Didel… lc                  19.4       6.6      NA       4.6 NA     
#>  9 Slow lor… Nycti… carni Prima… NA                  11        NA        NA      13    0.0125
#> 10 Northern… Onych… carni Roden… lc                  14.5      NA        NA       9.5 NA     
#> 11 Tiger     Panth… carni Carni… en                  15.8      NA        NA       8.2 NA     
#> 12 Jaguar    Panth… carni Carni… nt                  10.4      NA        NA      13.6  0.157 
#> 13 Lion      Panth… carni Carni… vu                  13.5      NA        NA      10.5 NA     
#> 14 Caspian … Phoca  carni Carni… vu                   3.5       0.4      NA      20.5 NA     
#> 15 Common p… Phoco… carni Cetac… vu                   5.6      NA        NA      18.4 NA     
#> 16 Bottle-n… Tursi… carni Cetac… NA                   5.2      NA        NA      18.8 NA     
#> 17 Genet     Genet… carni Carni… NA                   6.3       1.3      NA      17.7  0.0175
#> 18 Arctic f… Vulpes carni Carni… NA                  12.5      NA        NA      11.5  0.0445
#> 19 Red fox   Vulpes carni Carni… NA                   9.8       2.4       0.35   14.2  0.0504
#> # … with 1 more variable: bodywt <dbl>
#> 
#> $herbi
#> # A tibble: 32 × 11
#>    name     genus  vore  order   conservation sleep_total sleep_rem sleep_cycle awake brainwt
#>    <chr>    <chr>  <chr> <chr>   <chr>              <dbl>     <dbl>       <dbl> <dbl>   <dbl>
#>  1 Mountai… Aplod… herbi Rodent… nt                  14.4       2.4      NA       9.6 NA     
#>  2 Cow      Bos    herbi Artiod… domesticated         4         0.7       0.667  20    0.423 
#>  3 Three-t… Brady… herbi Pilosa  NA                  14.4       2.2       0.767   9.6 NA     
#>  4 Roe deer Capre… herbi Artiod… lc                   3        NA        NA      21    0.0982
#>  5 Goat     Capri  herbi Artiod… lc                   5.3       0.6      NA      18.7  0.115 
#>  6 Guinea … Cavis  herbi Rodent… domesticated         9.4       0.8       0.217  14.6  0.0055
#>  7 Chinchi… Chinc… herbi Rodent… domesticated        12.5       1.5       0.117  11.5  0.0064
#>  8 Tree hy… Dendr… herbi Hyraco… lc                   5.3       0.5      NA      18.7  0.0123
#>  9 Asian e… Eleph… herbi Probos… en                   3.9      NA        NA      20.1  4.60  
#> 10 Horse    Equus  herbi Periss… domesticated         2.9       0.6       1      21.1  0.655 
#> # … with 22 more rows, and 1 more variable: bodywt <dbl>
#> 
#> $insecti
#> # A tibble: 5 × 11
#>   name  genus vore  order conservation sleep_total sleep_rem sleep_cycle awake brainwt bodywt
#>   <chr> <chr> <chr> <chr> <chr>              <dbl>     <dbl>       <dbl> <dbl>   <dbl>  <dbl>
#> 1 Big … Epte… inse… Chir… lc                  19.7       3.9       0.117   4.3 0.0003   0.023
#> 2 Litt… Myot… inse… Chir… NA                  19.9       2         0.2     4.1 0.00025  0.01 
#> 3 Gian… Prio… inse… Cing… en                  18.1       6.1      NA       5.9 0.081   60    
#> 4 East… Scal… inse… Sori… lc                   8.4       2.1       0.167  15.6 0.0012   0.075
#> 5 Shor… Tach… inse… Mono… NA                   8.6      NA        NA      15.4 0.025    4.5  
#> 
#> $omni
#> # A tibble: 20 × 11
#>    name     genus  vore  order  conservation sleep_total sleep_rem sleep_cycle awake  brainwt
#>    <chr>    <chr>  <chr> <chr>  <chr>              <dbl>     <dbl>       <dbl> <dbl>    <dbl>
#>  1 Owl mon… Aotus  omni  Prima… NA                  17         1.8      NA       7    0.0155 
#>  2 Greater… Blari… omni  Soric… lc                  14.9       2.3       0.133   9.1  0.00029
#>  3 Grivet   Cerco… omni  Prima… lc                  10         0.7      NA      14   NA      
#>  4 Star-no… Condy… omni  Soric… lc                  10.3       2.2      NA      13.7  0.001  
#>  5 African… Crice… omni  Roden… NA                   8.3       2        NA      15.7  0.0066 
#>  6 Lesser … Crypt… omni  Soric… lc                   9.1       1.4       0.15   14.9  0.00014
#>  7 North A… Didel… omni  Didel… lc                  18         4.9       0.333   6    0.0063 
#>  8 Europea… Erina… omni  Erina… lc                  10.1       3.5       0.283  13.9  0.0035 
#>  9 Patas m… Eryth… omni  Prima… lc                  10.9       1.1      NA      13.1  0.115  
#> 10 Galago   Galago omni  Prima… NA                   9.8       1.1       0.55   14.2  0.005  
#> 11 Human    Homo   omni  Prima… NA                   8         1.9       1.5    16    1.32   
#> 12 Macaque  Macaca omni  Prima… NA                  10.1       1.2       0.75   13.9  0.179  
#> 13 Chimpan… Pan    omni  Prima… NA                   9.7       1.4       1.42   14.3  0.44   
#> 14 Baboon   Papio  omni  Prima… NA                   9.4       1         0.667  14.6  0.18   
#> 15 Potto    Perod… omni  Prima… lc                  11        NA        NA      13   NA      
#> 16 African… Rhabd… omni  Roden… NA                   8.7      NA        NA      15.3 NA      
#> 17 Squirre… Saimi… omni  Prima… NA                   9.6       1.4      NA      14.4  0.02   
#> 18 Pig      Sus    omni  Artio… domesticated         9.1       2.4       0.5    14.9  0.18   
#> 19 Tenrec   Tenrec omni  Afros… NA                  15.6       2.3      NA       8.4  0.0026 
#> 20 Tree sh… Tupaia omni  Scand… NA                   8.9       2.6       0.233  15.1  0.0025 
#> # … with 1 more variable: bodywt <dbl>
#> 
# }