Moral judgments about third-party moral behavior.

Moral judgments about third-party moral behavior.

intent_morality

Format

A data frame with 4016 rows and 8 variables

id. Participant id.

gender. Participant's gender.

item. Which story/vignette participants read for a given condition.

harm. What kind of harm was involved in the item.

belief. What kind of belief the actor had (neutral or negative/harmful).

outcome. What kind of outcome the actor caused (neutral or negative/harmful).

condition. Type of harm, composed of belif and outcome.

question. Type of moral judgment asked (wrongess or punishment).

rating. Moral judgment rating on a scale of 1 to 7.

Source

https://www.nature.com/articles/s41598-017-05299-9

Details

This dataset contains data from a recent study about how people judge behavior of others when they unintentionally or intentionally cause harm to others.

Participants responded to four different vignettes that contains four different types of conditions-

accidental harm. neutral belief, harmful/negative outcome

intentional harm. harmful/negative belief, harmful/negative outcome

attempted harm. harmful/negative belief, neutral outcome

neutral harm. neutral belief, neutral outcome

Additionally, participants saw one of the four variants for each of the four items. Each of the item had a different type of harm.

Examples

dim(intent_morality)

#> [1] 4016    9

head(intent_morality)

#> # A tibble: 6 x 9
#>      id gender item  harm      belief  outcome condition question  rating
#>   <int> <fct>  <fct> <fct>     <fct>   <fct>   <fct>     <fct>      <int>
#> 1     2 female Grace Poisoning neutral neutral neutral   wrongness      4
#> 2    11 female Grace Poisoning neutral neutral neutral   wrongness      1
#> 3    16 male   Grace Poisoning neutral neutral neutral   wrongness      1
#> 4    17 male   Grace Poisoning neutral neutral neutral   wrongness      1
#> 5    20 male   Grace Poisoning neutral neutral neutral   wrongness      2
#> 6    22 male   Grace Poisoning neutral neutral neutral   wrongness      1