Standardize a dataframe with effect sizes for aov, lm, aovlist, etc. objects.

Standardize a dataframe with effect sizes for aov, lm, aovlist, etc. objects.

lm_effsize_standardizer(object, effsize = "eta", partial = TRUE,
  conf.level = 0.95, nboot = 500)

Arguments

object
The linear model object (can be of class lm, aov, anova, or aovlist).

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega".

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class lm, aov, anova, or aovlist.

conf.level
Numeric specifying Level of confidence for the confidence interval (Default: 0.95).

nboot
Number of bootstrap samples for confidence intervals for partial eta-squared and omega-squared (Default: 500).

Examples

ggstatsplot:::lm_effsize_standardizer(
  stats::lm(formula = brainwt ~ vore, data = ggplot2::msleep),
  partial = FALSE,
  conf.level = 0.99,
  nboot = 50
)

#> # A tibble: 1 x 8
#>   term  F.value   df1   df2 p.value estimate conf.low conf.high
#>   <chr>   <dbl> <int> <int>   <dbl>    <dbl>    <dbl>     <dbl>
#> 1 vore     1.06     3    47   0.375   0.0633 0.000406     0.156