Converts long-format dataframe to wide-format dataframe

This conversion is helpful mostly for repeated measures design.

long_to_wide_converter(data, x, y, paired = TRUE)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

paired
A logical that indicates whether the design was repeated measures (within-subjects) or between-subjects (default: TRUE).

Examples

ggstatsplot:::long_to_wide_converter(
  data = iris_long,
  x = condition,
  y = value,
  paired = TRUE
)

#> # A tibble: 150 x 5
#>    rowid Petal.Length Petal.Width Sepal.Length Sepal.Width
#>    <int>        <dbl>       <dbl>        <dbl>       <dbl>
#>  1     1          1.4         0.2          5.1         3.5
#>  2     2          1.4         0.2          4.9         3  
#>  3     3          1.3         0.2          4.7         3.2
#>  4     4          1.5         0.2          4.6         3.1
#>  5     5          1.4         0.2          5           3.6
#>  6     6          1.7         0.4          5.4         3.9
#>  7     7          1.4         0.3          4.6         3.4
#>  8     8          1.5         0.2          5           3.4
#>  9     9          1.4         0.2          4.4         2.9
#> 10    10          1.5         0.1          4.9         3.1
#> # ... with 140 more rows