Adding a column to dataframe describing outlier status.

This function is mostly helpful for internal operations of some of the functions in this package.

outlier_df(data, x, y, outlier.label, outlier.coef = 1.5, ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

outlier.label
Label to put on the outliers that have been tagged.

outlier.coef
Coefficient for outlier detection using Tukey's method. With Tukey's method, outliers are below (1st Quartile) or above (3rd Quartile) outlier.coef times the Inter-Quartile Range (IQR) (Default: 1.5).

...
Additional arguments.

Examples

# adding column for outlier and a label for that outlier
ggstatsplot::outlier_df(
  data = morley,
  x = Expt,
  y = Speed,
  outlier.label = Run,
  outlier.coef = 2
) %>%
  dplyr::arrange(outlier)

#> # A tibble: 100 x 5
#>     Expt   Run Speed isanoutlier outlier
#>    <int> <int> <int> <lgl>         <int>
#>  1     3     5   720 TRUE              5
#>  2     3     6   720 TRUE              6
#>  3     3     7   620 TRUE              7
#>  4     3     9   970 TRUE              9
#>  5     1     1   850 FALSE            NA
#>  6     1     2   740 FALSE            NA
#>  7     1     3   900 FALSE            NA
#>  8     1     4  1070 FALSE            NA
#>  9     1     5   930 FALSE            NA
#> 10     1     6   850 FALSE            NA
#> # ... with 90 more rows