Pairwise comparison tests

Calculate pairwise comparisons between group levels with corrections for multiple testing.

pairwise_p(data, x, y, type = "parametric", tr = 0.1, paired = FALSE,
  var.equal = FALSE, p.adjust.method = "holm", k = 2,
  messages = TRUE, ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

type
Type of statistic expected ("parametric" or "nonparametric" or "robust" or "bayes").Corresponding abbreviations are also accepted: "p" (for parametric), "np" (nonparametric), "r" (robust), or "bf"resp.

tr
Trim level for the mean when carrying out robust tests. If you get error stating "Standard error cannot be computed because of Winsorized variance of 0 (e.g., due to ties). Try to decrease the trimming level.", try to play around with the value of tr, which is by default set to 0.1. Lowering the value might help.

paired
a logical indicating whether you want a paired t-test.

var.equal
a logical variable indicating whether to treat the variances in the samples as equal. If TRUE, then a simple F test for the equality of means in a one-way analysis of variance is performed. If FALSE, an approximate method of Welch (1951) is used, which generalizes the commonly known 2-sample Welch test to the case of arbitrarily many samples.

p.adjust.method
Adjustment method for p-values for multiple comparisons. Possible methods are: "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

...
Additional arguments.

See also

Examples

# time consuming, so not run on `CRAN` machines

# show all columns in a tibble
options(tibble.width = Inf)

# for reproducibility
set.seed(123)

#------------------- between-subjects design ----------------------------

# parametric
# if `var.equal = TRUE`, then Student's *t*-test will be run
ggstatsplot::pairwise_p(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  type = "p",
  var.equal = TRUE,
  paired = FALSE,
  p.adjust.method = "bonferroni"
)

#> Note: The parametric pairwise multiple comparisons test used-
#>  Student's t-test.
#>  Adjustment method for p-values: bonferroni

#> 

#> # A tibble: 6 x 6
#>   group1  group2  mean.difference p.value significance p.value.label
#>   <chr>   <chr>             <dbl>   <dbl> <chr>        <chr>        
#> 1 carni   herbi            0.542    1     ns           p = 1.000    
#> 2 carni   insecti         -0.0577   1     ns           p = 1.000    
#> 3 carni   omni             0.0665   1     ns           p = 1.000    
#> 4 herbi   insecti         -0.600    1     ns           p = 1.000    
#> 5 herbi   omni            -0.476    0.979 ns           p = 0.979    
#> 6 insecti omni             0.124    1     ns           p = 1.000    

# if `var.equal = FALSE`, then Games-Howell test will be run
ggstatsplot::pairwise_p(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  type = "p",
  var.equal = FALSE,
  paired = FALSE,
  p.adjust.method = "bonferroni"
)

#> Note: The parametric pairwise multiple comparisons test used-
#>  Games-Howell test.
#>  Adjustment method for p-values: bonferroni

#> 

#> # A tibble: 6 x 9
#>   group1 group2  mean.difference    se t.value    df p.value significance
#>   <chr>  <chr>             <dbl> <dbl>   <dbl> <dbl>   <dbl> <chr>       
#> 1 omni   herbi             0.476 0.255   1.32   20.9       1 ns          
#> 2 omni   carni            -0.066 0.061   0.774  21.1       1 ns          
#> 3 omni   insecti          -0.124 0.057   1.55   17.2       1 ns          
#> 4 herbi  carni            -0.542 0.25    1.54   19.4       1 ns          
#> 5 herbi  insecti          -0.6   0.249   1.70   19.1       1 ns          
#> 6 carni  insecti          -0.058 0.027   1.53   10.7       1 ns          
#>   p.value.label
#>   <chr>        
#> 1 p = 1.000    
#> 2 p = 1.000    
#> 3 p = 1.000    
#> 4 p = 1.000    
#> 5 p = 1.000    
#> 6 p = 1.000    

# non-parametric
ggstatsplot::pairwise_p(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  type = "np",
  paired = FALSE,
  p.adjust.method = "none"
)

#> Note: The nonparametric pairwise multiple comparisons test used-
#>  Dwass-Steel-Crichtlow-Fligner test.
#>  Adjustment method for p-values: none

#> 

#> # A tibble: 6 x 6
#>   group1  group2       W p.value significance p.value.label
#>   <chr>   <chr>    <dbl>   <dbl> <chr>        <chr>        
#> 1 carni   herbi   -0.8    0.572  ns           p = 0.572    
#> 2 carni   insecti -2.36   0.0956 ns           p = 0.096    
#> 3 carni   omni    -1.72   0.225  ns           p = 0.225    
#> 4 herbi   insecti -2.40   0.0895 ns           p = 0.089    
#> 5 herbi   omni    -0.948  0.503  ns           p = 0.503    
#> 6 insecti omni     1.61   0.256  ns           p = 0.256    

# robust
ggstatsplot::pairwise_p(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  type = "r",
  paired = FALSE,
  p.adjust.method = "fdr"
)

#> Note: The robust pairwise multiple comparisons test used-
#>  Yuen's trimmed means comparisons test.
#>  Adjustment method for p-values: fdr

#> 

#> # A tibble: 6 x 8
#>   group1  group2    psihat conf.low conf.high p.value significance p.value.label
#>   <chr>   <chr>      <dbl>    <dbl>     <dbl>   <dbl> <chr>        <chr>        
#> 1 insecti omni    -0.0556   -0.184     0.0728   0.969 ns           p = 0.969    
#> 2 carni   herbi   -0.0530   -0.274     0.168    0.969 ns           p = 0.969    
#> 3 carni   omni     0.00210  -0.151     0.155    0.969 ns           p = 0.969    
#> 4 herbi   omni     0.0551   -0.173     0.283    0.969 ns           p = 0.969    
#> 5 carni   insecti  0.0577   -0.0609    0.176    0.969 ns           p = 0.969    
#> 6 herbi   insecti  0.111    -0.0983    0.320    0.969 ns           p = 0.969    

#------------------- within-subjects design ----------------------------

set.seed(123)
library(jmv)
data("bugs", package = "jmv")

# converting to long format
bugs_long <- bugs %>%
  tibble::as_tibble(.) %>%
  tidyr::gather(., key, value, LDLF:HDHF)

# parametric
ggstatsplot::pairwise_p(
  data = bugs_long,
  x = key,
  y = value,
  type = "p",
  paired = TRUE,
  p.adjust.method = "BH"
)

#> Note: The parametric pairwise multiple comparisons test used-
#>  Student's t-test.
#>  Adjustment method for p-values: BH

#> 

#> # A tibble: 6 x 6
#>   group1 group2 mean.difference   p.value significance p.value.label
#>   <chr>  <chr>            <dbl>     <dbl> <chr>        <chr>        
#> 1 HDHF   HDLF            -1.11  10.00e- 4 ***          p = 0.001    
#> 2 HDHF   LDHF            -0.474  7.07e- 2 ns           p = 0.071    
#> 3 HDHF   LDLF            -2.14   7.64e-12 ***          p <= 0.001   
#> 4 HDLF   LDHF             0.637  5.47e- 2 ns           p = 0.055    
#> 5 HDLF   LDLF            -1.03   1.39e- 3 **           p = 0.001    
#> 6 LDHF   LDLF            -1.66   6.67e- 9 ***          p <= 0.001   

# non-parametric
ggstatsplot::pairwise_p(
  data = bugs_long,
  x = key,
  y = value,
  type = "np",
  paired = TRUE,
  p.adjust.method = "BY"
)

#> Note: The nonparametric pairwise multiple comparisons test used-
#>  Durbin-Conover test.
#>  Adjustment method for p-values: BY

#> 

#> # A tibble: 6 x 6
#>   group1 group2 statistic  p.value significance p.value.label
#>   <chr>  <chr>      <dbl>    <dbl> <chr>        <chr>        
#> 1 HDHF   HDLF        4.78 1.44e- 5 ***          p <= 0.001   
#> 2 HDHF   LDHF        2.44 4.47e- 2 *            p = 0.045    
#> 3 HDHF   LDLF        8.01 5.45e-13 ***          p <= 0.001   
#> 4 HDLF   LDHF        2.34 4.96e- 2 *            p = 0.050    
#> 5 HDLF   LDLF        3.23 5.05e- 3 **           p = 0.005    
#> 6 LDHF   LDLF        5.57 4.64e- 7 ***          p <= 0.001   

# robust
ggstatsplot::pairwise_p(
  data = bugs_long,
  x = key,
  y = value,
  type = "r",
  paired = TRUE,
  p.adjust.method = "hommel"
)

#> Note: The robust pairwise multiple comparisons test used-
#>  Yuen's trimmed means comparisons test.
#>  Adjustment method for p-values: hommel

#> 

#> # A tibble: 6 x 8
#>   group1 group2 psihat conf.low conf.high  p.value significance p.value.label
#>   <chr>  <chr>   <dbl>    <dbl>     <dbl>    <dbl> <chr>        <chr>        
#> 1 HDLF   LDHF   -0.701  -1.71       0.303 6.20e- 2 ns           p = 0.062    
#> 2 HDHF   LDHF    0.5    -0.188      1.19  6.20e- 2 ns           p = 0.062    
#> 3 HDLF   LDLF    0.938   0.0694     1.81  1.36e- 2 *            p = 0.014    
#> 4 HDHF   HDLF    1.16    0.318      2.00  1.49e- 3 **           p = 0.001    
#> 5 LDHF   LDLF    1.54    0.810      2.27  1.16e- 6 ***          p <= 0.001   
#> 6 HDHF   LDLF    2.10    1.37       2.82  1.79e-10 ***          p <= 0.001