Creating a new character type column with significance labels

This function will add a new column to a dataframe containing p-values

signif_column(data = NULL, p)

Arguments

data
Data frame from which variables specified are preferentially to be taken.

p
The column containing p-values.

Value

Returns the originally entered object (either a vector or a dataframe) in tibble format with an additional column corresponding to statistical significance.

See also

Examples

# vector as input
signif_column(p = c(0.05, 0.1, 1, 0.00001, 0.001, 0.01))

#> # A tibble: 6 x 2
#>         p significance
#>     <dbl> <chr>       
#> 1 0.05    ns          
#> 2 0.1     ns          
#> 3 1       ns          
#> 4 0.00001 ***         
#> 5 0.001   **          
#> 6 0.01    *           

# dataframe as input
# preparing a newdataframe
df <- tibble(
  x = 1:5,
  y = 1,
  p.value = c(0.1, 0.5, 0.00001, 0.05, 0.01)
)

signif_column(data = df, p = p.value)

#> # A tibble: 5 x 4
#>       x     y p.value significance
#>   <int> <dbl>   <dbl> <chr>       
#> 1     1     1 0.1     ns          
#> 2     2     1 0.5     ns          
#> 3     3     1 0.00001 ***         
#> 4     4     1 0.05    ns          
#> 5     5     1 0.01    *