Making text subtitle for parametric ANOVA.

Making text subtitle for parametric ANOVA.

subtitle_anova_parametric(data, x, y, paired = FALSE,
  effsize.type = "unbiased", partial = TRUE, conf.level = 0.95,
  nboot = 100, var.equal = FALSE, sphericity.correction = TRUE,
  k = 2, stat.title = NULL, messages = TRUE, ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

paired
Logical that decides whether the design is repeated measures/within-subjects (in which case one-way Friedman Rank Sum Test will be carried out) or between-subjects (in which case one-way Kruskal–Wallis H test will be carried out). The default is FALSE.

effsize.type
Type of effect size needed for parametric tests. The argument can be "biased" ("d" for Cohen's d for t-test; "partial_eta" for partial eta-squared for anova) or "unbiased" ("g" Hedge's g for t-test; "partial_omega" for partial omega-squared for anova)).

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class lm, aov, anova, or aovlist.

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% lower and upper confidence intervals (0.95).

nboot
Number of bootstrap samples for computing confidence interval for the effect size (Default: 100).

var.equal
a logical variable indicating whether to treat the variances in the samples as equal. If TRUE, then a simple F test for the equality of means in a one-way analysis of variance is performed. If FALSE, an approximate method of Welch (1951) is used, which generalizes the commonly known 2-sample Welch test to the case of arbitrarily many samples.

sphericity.correction
Logical that decides whether to apply correction to account for violation of sphericity in a repeated measures design ANOVA (Default: TRUE).

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

stat.title
A character describing the test being run, which will be added as a prefix in the subtitle. The default is NULL. An example of a stat.title argument will be something like "Student's t-test: ".

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

...
Additional arguments.

Note

For repeated measures designs (paired = TRUE), only omega-squared and partial eta-squared effect sizes are supported.

Examples

# for reproducibility
set.seed(123)
library(ggstatsplot)

# -------------------- between-subjects ------------------------------

# with defaults
ggstatsplot::subtitle_anova_parametric(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem,
  paired = FALSE,
  k = 3
)

#> Note: 95% CI for effect size estimate was computed with 100 bootstrap samples.
#> 

#> 

#> paste(NULL, italic("F"), "(", "3", ",", "11.101", ") = ", "2.633", 
#>     ", ", italic("p"), " = ", "0.102", ", ", omega["p"]^2, " = ", 
#>     "0.144", ", CI"["95%"], " [", "-0.025", ", ", "0.421", "]", 
#>     ", ", italic("n"), " = ", 56L)

# modifying the defaults
ggstatsplot::subtitle_anova_parametric(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem,
  paired = FALSE,
  effsize.type = "biased",
  partial = FALSE,
  var.equal = TRUE,
  nboot = 10
)

#> Note: 95% CI for effect size estimate was computed with 10 bootstrap samples.
#> 

#> 

#> paste(NULL, italic("F"), "(", "3", ",", "52", ") = ", "4.14", 
#>     ", ", italic("p"), " = ", "0.011", ", ", eta^2, " = ", "0.19", 
#>     ", CI"["95%"], " [", "0.04", ", ", "0.52", "]", ", ", italic("n"), 
#>     " = ", 56L)

# -------------------- repeated measures ------------------------------

ggstatsplot::subtitle_anova_parametric(
  data = iris_long,
  x = condition,
  y = value,
  paired = TRUE,
  k = 4,
  nboot = 10
)

#> Note: 95% CI for effect size estimate was computed with 10 bootstrap samples.
#> 

#> 

#> paste(NULL, italic("F"), "(", "1.1491", ",", "171.2174", ") = ", 
#>     "776.3182", ", ", italic("p"), " = ", "< 0.001", ", ", omega^2, 
#>     " = ", "0.7074", ", CI"["95%"], " [", "0.7675", ", ", "0.8181", 
#>     "]", ", ", italic("n"), " = ", 150L)