Making text subtitle for the correlation test.

Making text subtitle for the correlation test.

subtitle_ggscatterstats(data, x, y, nboot = 100, beta = 0.1,
  type = "pearson", conf.level = 0.95, conf.type = "norm",
  messages = TRUE, k = 3)

Arguments

data
Dataframe from which variables specified are preferentially to be taken.

x
A vector containing the explanatory variable.

y
The response - a vector of length the number of rows of x.

nboot
Number of bootstrap samples for computing effect size (Default: 100).

beta
bending constant (Default: 0.1). For more, see ?WRS2::pbcor.

type
Type of association between paired samples required (""parametric": Pearson's product moment correlation coefficient" or ""nonparametric": Spearman's rho" or ""robust": percentage bend correlation coefficient"). Corresponding abbreviations are also accepted: "p" (for parametric/pearson's), "np" (nonparametric/spearman), "r" (robust), resp.

conf.level
Scalar between 0 and 1. If NULL, the defaults return 95 lower and upper confidence intervals (0.95).

conf.type
A vector of character strings representing the type of intervals required. The value should be any subset of the values "norm", "basic", "perc", "bca". For more, see ?boot::boot.ci.

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

k
Number of digits after decimal point (should be an integer) (Default: k = 3).