Making text subtitle for Proportion Test (N Outcomes), a chi-squared Goodness of fit test.

Making text subtitle for Proportion Test (N Outcomes), a chi-squared Goodness of fit test.

subtitle_onesample_proptest(data, main, counts = NULL, ratio = NULL,
  legend.title = NULL, k = 3)

Arguments

data
The data as a data frame (matrix or tables will not be accepted).

main
The variable to use as the rows in the contingency table.

counts
A string naming a variable in data containing counts, or NULL if each row represents a single observation (Default).

ratio
A vector of numbers: the expected proportions for the proportion test. Default is NULL, which means if there are two levels ratio = c(1,1), etc.

legend.title
Title text for the legend.

k
Number of digits after decimal point (should be an integer) (Default: k = 3).

Examples

# with counts
library(jmv)

subtitle_onesample_proptest(
  data = as.data.frame(HairEyeColor),
  main = Eye,
  counts = Freq
)

#> paste(italic(chi)^2, "(", 3, ") = ", "133.473", ", ", italic("p"), 
#>     " = ", "< 0.001", ", ", italic("n"), " = ", 592L)

# in case no variation, only sample size will be shown
subtitle_onesample_proptest(
  data = cbind.data.frame(x = rep("a", 10)),
  main = x
)

#> Warning:  Proportion test will not be run because it requires x to have at least 2 levels with non-zero frequencies.

#> 

#> paste(italic("n"), " = ", 10L)