Prepare labels with statistic for ggcoefstats function.

Creates text labels for ggcoefstats plot with the appropriate statistic (t, z, or f) displayed in the label.

tfz_labeller(tidy_df, glance_df = NULL, statistic, effsize = "eta",
  partial = TRUE, k = 2)

Arguments

tidy_df
Tidy dataframe from broomExtra::tidy.

glance_df
Glance model summary dataframe from broom::glance (default: NULL). This is optional argument. If provide, the glance summary will be used to write caption for the final plot.

statistic
Which statistic is to be displayed (either "t" or "f"or "z") in the label. This is especially important if the x argument in ggcoefstats is a dataframe in which case the function wouldn't know what kind of model it is dealing with.

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega". This argument is relevant only for models objects of class aov, anova, and aovlist.

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class aov, anova, or aovlist.

k
Number of decimal places expected for results displayed in labels (Default : k = 2).