Default theme used for correlation matrix

Default theme used for correlation matrix

theme_corrmat()

Value

A ggplot2 object with the theme_mprl theme.

Note

This is a helper function used internally in the package and not exported. In case you want to use it, you can do so by ggstatsplot:::theme_corrmat. Note that it is ::: and not ::.