Default theme used in all ggstatsplot package plots

Common theme used across all plots generated in ggstatsplot and assumed by the author to be aesthetically pleasing to the user/reader.

theme_mprl(ggtheme = ggplot2::theme_bw(), ggstatsplot.layer = TRUE)

Arguments

ggtheme
A function, ggplot2 theme name. Default value is ggplot2::theme_bw(). Any of the ggplot2 themes, or themes from extension packages are allowed (e.g., ggthemes::theme_economist(), hrbrthemes::theme_ipsum_ps(), ggthemes::theme_fivethirtyeight()).

ggstatsplot.layer
Logical that decides whether theme_ggstatsplot theme elements are to be displayed along with the selected ggtheme (Default: TRUE).

Note

This is just another name for the function theme_ggstatsplot().

References

https://indrajeetpatil.github.io/ggstatsplot/articles/theme_ggstatsplot.html