Default theme used for pie chart

Default theme used for pie chart

theme_pie(ggtheme = ggplot2::theme_bw(), ggstatsplot.layer = TRUE)

Arguments

ggtheme
A function, ggplot2 theme name. Default value is ggplot2::theme_bw(). Any of the ggplot2 themes, or themes from extension packages are allowed (e.g., ggthemes::theme_economist(), hrbrthemes::theme_ipsum_ps(), ggthemes::theme_fivethirtyeight()).

ggstatsplot.layer
Logical that decides whether theme_ggstatsplot theme elements are to be displayed along with the selected ggtheme (Default: TRUE).

Value

A ggplot2 object with the theme_ggstatsplot theme.

Note

This is a helper function used internally in the package and not exported. In case you want to use it, you can do so by ggstatsplot:::theme_pie. Note that it is ::: and not ::.