Making aesthetic modifications to the plot.

Making aesthetic modifications to the plot.

aesthetic_addon(plot, x, xlab = NULL, ylab = NULL, title = NULL,
  subtitle = NULL, caption = NULL, ggtheme = ggplot2::theme_bw(),
  ggstatsplot.layer = TRUE, package = "RColorBrewer",
  palette = "Dark2", direction = 1, ggplot.component = NULL, ...)

Arguments

plot
Plot to be aesthetically modified.

x
A numeric vector for x axis.

xlab
Labels for x and y axis variables. If NULL (default), variable names for x and y will be used.

ylab
Labels for x and y axis variables. If NULL (default), variable names for x and y will be used.

title
The text for the plot title.

subtitle
The text for the plot subtitle. Will work only if results.subtitle = FALSE.

caption
The text for the plot caption.

ggtheme
A function, ggplot2 theme name. Default value is ggplot2::theme_bw(). Any of the ggplot2 themes, or themes from extension packages are allowed (e.g., ggthemes::theme_fivethirtyeight(), hrbrthemes::theme_ipsum_ps(), etc.).

ggstatsplot.layer
Logical that decides whether theme_ggstatsplot theme elements are to be displayed along with the selected ggtheme (Default: TRUE).

package
Name of package from which the palette is desired as string or symbol.

palette
If a character string (e.g., "Set1"), will use that named palette. If a number, will index into the list of palettes of appropriate type. Default palette is "Dark2".

direction
Either 1 or -1. If -1 the palette will be reversed.

ggplot.component
A ggplot component to be added to the plot prepared by ggstatsplot. This argument is primarily helpful for grouped_ variant of the current function. Default is NULL. The argument should be entered as a function. If the given function has an argument axes.range.restrict and if it has been set to TRUE, the added ggplot component might not work as expected.

...
Additional arguments.