Display homogeneity of variance test as a message

A note to the user about the validity of assumptions for the default linear model.

bartlett_message(data, x, y, lab = NULL, k = 2, output = "message",
  ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

x
The grouping variable from the dataframe data.

y
The response (a.k.a. outcome or dependent) variable from the dataframe data.

lab
A character describing label for the variable. If NULL, variable name will be used.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

output
What output is desired: "message" (default) or "stats" (or "tidy") objects.

...
Additional arguments (ignored).

Value

A list of class "htest" containing the following components:

statistic

Bartlett's K-squared test statistic.

parameter

the degrees of freedom of the approximate chi-squared distribution of the test statistic.

p.value

the p-value of the test.

method

the character string "Bartlett test of homogeneity of variances".

data.name

a character string giving the names of the data.

See also

Examples

# getting message
ggstatsplot::bartlett_message(
  data = iris,
  x = Species,
  y = Sepal.Length,
  lab = "Iris Species"
)

#> Note: Bartlett's test for homogeneity of variances for factor Iris Species: p-value = < 0.001

#> 

# getting results from the test
ggstatsplot::bartlett_message(
  data = mtcars,
  x = am,
  y = wt,
  output = "tidy"
)

#> # A tibble: 1 x 4
#>   statistic p.value parameter method                                   
#>       <dbl>   <dbl>     <dbl> <chr>                                    
#> 1     0.715   0.398         1 Bartlett test of homogeneity of variances