Bayes factor message for random-effects meta-analysis

Bayes factor message for random-effects meta-analysis

bf_meta_message(data, k = 2, d = "norm", d.par = c(mean = 0, sd =
  0.3), tau = "halfcauchy", tau.par = c(scale = 0.5), iter = 10000,
  summarize = "stan", caption = NULL, messages = TRUE, ...)

Arguments

data
data frame containing the variables for effect size y, standard error SE, labels, and moderators per study.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

d
the prior distribution of the average effect size \(d\) specified either as the type of family (e.g., "norm") or via prior

d.par
prior parameters for \(d\) (only used if d specifies the type of family)

tau
the prior distribution of the between-study heterogeneity \(\tau\) specified either as a character value (e.g., "halfcauchy") or via prior

tau.par
prior parameters for \(\tau\) (only used if tau specifies the type of family)

iter
number of MCMC iterations using Stan.

summarize
how to estimate parameter summaries (mean, median, SD, etc.): Either by numerical integration (summarize = "integrate") or based on MCMC/Stan samples (summarize = "stan").

caption
The text for the plot caption.

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

...
further arguments passed to rstan::sampling (see stanmodel-method-sampling). For instance: warmup=500, chains=4, control=list(adapt_delta=.95)).

Examples

# setup
set.seed(123)
library(metaBMA)

#> Loading required package: Rcpp

# creating a dataframe
(df <-
  structure(
    .Data = list(
      study = c("1", "2", "3", "4", "5"),
      estimate = c(
        0.382047603321706,
        0.780783111514665,
        0.425607573765058,
        0.558365541235078,
        0.956473848429961
      ),
      std.error = c(
        0.0465576338644502,
        0.0330218199731529,
        0.0362834986178494,
        0.0480571500648261,
        0.062215818388157
      )
    ),
    row.names = c(NA, -5L),
    class = c("tbl_df", "tbl", "data.frame")
  ))

#> # A tibble: 5 x 3
#>   study estimate std.error
#>   <chr>    <dbl>     <dbl>
#> 1 1        0.382    0.0466
#> 2 2        0.781    0.0330
#> 3 3        0.426    0.0363
#> 4 4        0.558    0.0481
#> 5 5        0.956    0.0622

# getting Bayes factor in favor of null hypothesis
ggstatsplot::bf_meta_message(
  data = df,
  k = 3,
  iter = 1500,
  messages = TRUE
)

#> ### Bayesian Random-Effects Meta-Analysis ### 
#>    Prior on d:      'norm' (mean=0, sd=0.3) with support on the interval [-Inf,Inf]. 
#>    Prior on tau:    't' (location=0, scale=0.5, nu=1) truncated to the interval [0,Inf]. 
#> 
#> # Bayes factors:
#>            (denominator)
#> (numerator) random_H0 random_H1
#>   random_H0       1.0    0.0686
#>   random_H1      14.6    1.0000
#> 
#> # Posterior summary statistics of random-effects model:
#>      mean    sd  2.5%   50% 97.5% hpd95_lower hpd95_upper n_eff  Rhat
#> d   0.492 0.157 0.104 0.522 0.733       0.154       0.761 557.1 1.002
#> tau 0.335 0.171 0.144 0.288 0.802       0.111       0.672 603.7 1.002

#> atop(displaystyle(NULL), expr = paste("In favor of null: ", "log"["e"], 
#>     "(BF"["01"], ") = ", "-2.680", ", ", italic("d")["mean"]^"posterior", 
#>     " = ", "0.492", ", CI"["95%"], " [", "0.154", ", ", "0.761", 
#>     "]"))