Split data frame into a list by grouping variable.

This function splits the data frame into a list, with the length of the list equal to the factor levels of the grouping variable.

Usage

.grouped_list(data, grouping.var = NULL)

Arguments

data

A data frame (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted. Additionally, grouped data frames from {dplyr} should be ungrouped before they are entered as data.

grouping.var

A single grouping variable.

Examples

ggstatsplot:::.grouped_list(ggplot2::msleep, grouping.var = vore)
#> $data
#> $data$carni
#> # A tibble: 19 × 11
#>    name  genus vore  order conse…¹ sleep…² sleep…³ sleep…⁴ awake brainwt  bodywt
#>    <chr> <chr> <chr> <chr> <chr>     <dbl>   <dbl>   <dbl> <dbl>   <dbl>   <dbl>
#>  1 Chee… Acin… carni Carn… lc         12.1    NA    NA      11.9 NA       50    
#>  2 Nort… Call… carni Carn… vu          8.7     1.4   0.383  15.3 NA       20.5  
#>  3 Dog   Canis carni Carn… domest…    10.1     2.9   0.333  13.9  0.07    14    
#>  4 Long… Dasy… carni Cing… lc         17.4     3.1   0.383   6.6  0.0108   3.5  
#>  5 Dome… Felis carni Carn… domest…    12.5     3.2   0.417  11.5  0.0256   3.3  
#>  6 Pilo… Glob… carni Ceta… cd          2.7     0.1  NA      21.4 NA      800    
#>  7 Gray… Hali… carni Carn… lc          6.2     1.5  NA      17.8  0.325   85    
#>  8 Thic… Lutr… carni Dide… lc         19.4     6.6  NA       4.6 NA        0.37 
#>  9 Slow… Nyct… carni Prim… NA         11      NA    NA      13    0.0125   1.4  
#> 10 Nort… Onyc… carni Rode… lc         14.5    NA    NA       9.5 NA        0.028
#> 11 Tiger Pant… carni Carn… en         15.8    NA    NA       8.2 NA      163.   
#> 12 Jagu… Pant… carni Carn… nt         10.4    NA    NA      13.6  0.157  100    
#> 13 Lion  Pant… carni Carn… vu         13.5    NA    NA      10.5 NA      161.   
#> 14 Casp… Phoca carni Carn… vu          3.5     0.4  NA      20.5 NA       86    
#> 15 Comm… Phoc… carni Ceta… vu          5.6    NA    NA      18.4 NA       53.2  
#> 16 Bott… Turs… carni Ceta… NA          5.2    NA    NA      18.8 NA      173.   
#> 17 Genet Gene… carni Carn… NA          6.3     1.3  NA      17.7  0.0175   2    
#> 18 Arct… Vulp… carni Carn… NA         12.5    NA    NA      11.5  0.0445   3.38 
#> 19 Red … Vulp… carni Carn… NA          9.8     2.4   0.35   14.2  0.0504   4.23 
#> # … with abbreviated variable names ¹​conservation, ²​sleep_total, ³​sleep_rem,
#> #   ⁴​sleep_cycle
#> 
#> $data$herbi
#> # A tibble: 32 × 11
#>    name  genus vore  order conse…¹ sleep…² sleep…³ sleep…⁴ awake brainwt  bodywt
#>    <chr> <chr> <chr> <chr> <chr>     <dbl>   <dbl>   <dbl> <dbl>   <dbl>   <dbl>
#>  1 Moun… Aplo… herbi Rode… nt         14.4     2.4  NA       9.6 NA      1.35e+0
#>  2 Cow   Bos   herbi Arti… domest…     4       0.7   0.667  20    0.423  6   e+2
#>  3 Thre… Brad… herbi Pilo… NA         14.4     2.2   0.767   9.6 NA      3.85e+0
#>  4 Roe … Capr… herbi Arti… lc          3      NA    NA      21    0.0982 1.48e+1
#>  5 Goat  Capri herbi Arti… lc          5.3     0.6  NA      18.7  0.115  3.35e+1
#>  6 Guin… Cavis herbi Rode… domest…     9.4     0.8   0.217  14.6  0.0055 7.28e-1
#>  7 Chin… Chin… herbi Rode… domest…    12.5     1.5   0.117  11.5  0.0064 4.2 e-1
#>  8 Tree… Dend… herbi Hyra… lc          5.3     0.5  NA      18.7  0.0123 2.95e+0
#>  9 Asia… Elep… herbi Prob… en          3.9    NA    NA      20.1  4.60   2.55e+3
#> 10 Horse Equus herbi Peri… domest…     2.9     0.6   1      21.1  0.655  5.21e+2
#> # … with 22 more rows, and abbreviated variable names ¹​conservation,
#> #   ²​sleep_total, ³​sleep_rem, ⁴​sleep_cycle
#> 
#> $data$insecti
#> # A tibble: 5 × 11
#>   name    genus vore  order conse…¹ sleep…² sleep…³ sleep…⁴ awake brainwt bodywt
#>   <chr>   <chr> <chr> <chr> <chr>     <dbl>   <dbl>   <dbl> <dbl>   <dbl>  <dbl>
#> 1 Big br… Epte… inse… Chir… lc         19.7     3.9   0.117   4.3 0.0003   0.023
#> 2 Little… Myot… inse… Chir… NA         19.9     2     0.2     4.1 0.00025  0.01 
#> 3 Giant … Prio… inse… Cing… en         18.1     6.1  NA       5.9 0.081   60    
#> 4 Easter… Scal… inse… Sori… lc          8.4     2.1   0.167  15.6 0.0012   0.075
#> 5 Short-… Tach… inse… Mono… NA          8.6    NA    NA      15.4 0.025    4.5  
#> # … with abbreviated variable names ¹​conservation, ²​sleep_total, ³​sleep_rem,
#> #   ⁴​sleep_cycle
#> 
#> $data$omni
#> # A tibble: 20 × 11
#>    name  genus vore  order conse…¹ sleep…² sleep…³ sleep…⁴ awake  brainwt bodywt
#>    <chr> <chr> <chr> <chr> <chr>     <dbl>   <dbl>   <dbl> <dbl>    <dbl>  <dbl>
#>  1 Owl … Aotus omni  Prim… NA         17       1.8  NA       7    0.0155   0.48 
#>  2 Grea… Blar… omni  Sori… lc         14.9     2.3   0.133   9.1  0.00029  0.019
#>  3 Griv… Cerc… omni  Prim… lc         10       0.7  NA      14   NA        4.75 
#>  4 Star… Cond… omni  Sori… lc         10.3     2.2  NA      13.7  0.001    0.06 
#>  5 Afri… Cric… omni  Rode… NA          8.3     2    NA      15.7  0.0066   1    
#>  6 Less… Cryp… omni  Sori… lc          9.1     1.4   0.15   14.9  0.00014  0.005
#>  7 Nort… Dide… omni  Dide… lc         18       4.9   0.333   6    0.0063   1.7  
#>  8 Euro… Erin… omni  Erin… lc         10.1     3.5   0.283  13.9  0.0035   0.77 
#>  9 Pata… Eryt… omni  Prim… lc         10.9     1.1  NA      13.1  0.115   10    
#> 10 Gala… Gala… omni  Prim… NA          9.8     1.1   0.55   14.2  0.005    0.2  
#> 11 Human Homo  omni  Prim… NA          8       1.9   1.5    16    1.32    62    
#> 12 Maca… Maca… omni  Prim… NA         10.1     1.2   0.75   13.9  0.179    6.8  
#> 13 Chim… Pan   omni  Prim… NA          9.7     1.4   1.42   14.3  0.44    52.2  
#> 14 Babo… Papio omni  Prim… NA          9.4     1     0.667  14.6  0.18    25.2  
#> 15 Potto Pero… omni  Prim… lc         11      NA    NA      13   NA        1.1  
#> 16 Afri… Rhab… omni  Rode… NA          8.7    NA    NA      15.3 NA        0.044
#> 17 Squi… Saim… omni  Prim… NA          9.6     1.4  NA      14.4  0.02     0.743
#> 18 Pig   Sus   omni  Arti… domest…     9.1     2.4   0.5    14.9  0.18    86.2  
#> 19 Tenr… Tenr… omni  Afro… NA         15.6     2.3  NA       8.4  0.0026   0.9  
#> 20 Tree… Tupa… omni  Scan… NA          8.9     2.6   0.233  15.1  0.0025   0.104
#> # … with abbreviated variable names ¹​conservation, ²​sleep_total, ³​sleep_rem,
#> #   ⁴​sleep_cycle
#> 
#> 
#> $title
#> [1] "carni"   "herbi"   "insecti" "omni"   
#>