Extracting dataframes with statistical details from {ggstatsplot}

Extracting dataframes with statistical details from {ggstatsplot}

extract_stats(p, ...)

Arguments

p
A plot from {ggstatsplot} package

...
Ignored

Value

A list of tibbles containing statistical analysis summaries.

Details

This is a convenience function to extract dataframes with statistical details that are used to create expressions displayed in {ggstatsplot} plots as subtitle and/or as caption. Note that all of this analysis is carried out by the statsExpressions package.

For more details about underlying tests and effect size estimates, see the following vignette: https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Note

Note that if you want to use this function with ggscatterstats, you will have to set marginal = FALSE to return an object of ggplot type.

Examples

# \donttest{
set.seed(123)
library(ggstatsplot)

# in case of group comparisons
p <- ggbetweenstats(mtcars, cyl, mpg)
extract_stats(p)
#> $subtitle_data
#> # A tibble: 1 x 13
#>   statistic    df df.error    p.value
#>       <dbl> <dbl>    <dbl>      <dbl>
#> 1      31.6     2     18.0 0.00000127
#>   method                                                   effectsize estimate
#>   <chr>                                                    <chr>         <dbl>
#> 1 One-way analysis of means (not assuming equal variances) Omega2        0.744
#>   conf.level conf.low conf.high conf.method conf.distribution expression
#>        <dbl>    <dbl>     <dbl> <chr>       <chr>             <list>    
#> 1       0.95    0.531         1 ncp         F                 <language>
#> 
#> $caption_data
#> # A tibble: 6 x 20
#>   term  estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>    <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 mu      20.5         0.95  19.3        21.9  1              0     
#> 2 cyl-4    5.90        0.95   4.11        7.52 1              0     
#> 3 cyl-6   -0.704       0.95  -2.64        1.06 0.780          0.390 
#> 4 cyl-8   -5.18        0.95  -6.76       -3.55 1              0     
#> 5 sig2    11.0         0.95   6.24       18.3  1              0     
#> 6 g_cyl    2.69        0.95   0.0911     18.7  1              0.0418
#>   prior.distribution prior.location prior.scale component       bf10
#>   <chr>                       <dbl>       <dbl> <chr>          <dbl>
#> 1 cauchy                          0       0.707 extra       3008850.
#> 2 cauchy                          0       0.707 conditional 3008850.
#> 3 cauchy                          0       0.707 conditional 3008850.
#> 4 cauchy                          0       0.707 conditional 3008850.
#> 5 cauchy                          0       0.707 extra       3008850.
#> 6 cauchy                          0       0.707 extra       3008850.
#>   method                          log_e_bf10    r2 std.dev r2.conf.level
#>   <chr>                                <dbl> <dbl>   <dbl>         <dbl>
#> 1 Bayes factors for linear models       14.9 0.714  0.0503          0.95
#> 2 Bayes factors for linear models       14.9 0.714  0.0503          0.95
#> 3 Bayes factors for linear models       14.9 0.714  0.0503          0.95
#> 4 Bayes factors for linear models       14.9 0.714  0.0503          0.95
#> 5 Bayes factors for linear models       14.9 0.714  0.0503          0.95
#> 6 Bayes factors for linear models       14.9 0.714  0.0503          0.95
#>   r2.conf.low r2.conf.high expression
#>         <dbl>        <dbl> <list>    
#> 1       0.574        0.788 <language>
#> 2       0.574        0.788 <language>
#> 3       0.574        0.788 <language>
#> 4       0.574        0.788 <language>
#> 5       0.574        0.788 <language>
#> 6       0.574        0.788 <language>
#> 
#> $pairwise_comparisons_data
#> # A tibble: 3 x 11
#>   group1 group2 statistic   p.value alternative method            distribution
#>   <chr>  <chr>      <dbl>     <dbl> <chr>       <chr>             <chr>       
#> 1 4      6          -6.67 0.00110   two.sided   Games-Howell test q           
#> 2 4      8         -10.7  0.0000140 two.sided   Games-Howell test q           
#> 3 6      8          -7.48 0.000257  two.sided   Games-Howell test q           
#>   p.adjustment test.details      p.value.adjustment
#>   <chr>        <chr>             <chr>             
#> 1 none         Games-Howell test Holm              
#> 2 none         Games-Howell test Holm              
#> 3 none         Games-Howell test Holm              
#>   label                                     
#>   <chr>                                     
#> 1 list(~italic(p)[Holm-corrected]==0.001)   
#> 2 list(~italic(p)[Holm-corrected]==1.4e-05) 
#> 3 list(~italic(p)[Holm-corrected]==2.57e-04)
#> 
#> $descriptive_data
#> NULL
#> 
#> $one_sample_data
#> NULL
#> 

# the exact details depend on the function
extract_stats(ggbarstats(mtcars, cyl, am))
#> $subtitle_data
#> # A tibble: 1 x 12
#>   statistic    df p.value method                     estimate conf.level
#>       <dbl> <int>   <dbl> <chr>                         <dbl>      <dbl>
#> 1      8.74     2  0.0126 Pearson's Chi-squared test    0.464       0.95
#>   conf.low conf.high effectsize        conf.method conf.distribution expression
#>      <dbl>     <dbl> <chr>             <chr>       <chr>             <list>    
#> 1        0         1 Cramer's V (adj.) ncp         chisq             <language>
#> 
#> $caption_data
#> # A tibble: 8 x 14
#>   term      estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>        <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 cell[1,1]    3.16        0.95    0.617     6.41      1               0
#> 2 cell[2,1]    3.97        0.95    1.04      7.65      1               0
#> 3 cell[3,1]   10.8         0.95    6.56     15.9       1               0
#> 4 cell[1,2]    7.44        0.95    3.50     11.8       1               0
#> 5 cell[2,2]    3.15        0.95    0.698     6.52      1               0
#> 6 cell[3,2]    2.29        0.95    0.409     5.38      1               0
#> 7 Ratio       NA          NA      NA        NA        NA              NA
#> 8 Cramers_v    0.476       0.95    0.205     0.703     1               0
#>   prior.distribution      prior.location prior.scale  bf10
#>   <chr>                            <dbl>       <dbl> <dbl>
#> 1 independent multinomial              0           1  15.1
#> 2 independent multinomial              0           1  15.1
#> 3 independent multinomial              0           1  15.1
#> 4 independent multinomial              0           1  15.1
#> 5 independent multinomial              0           1  15.1
#> 6 independent multinomial              0           1  15.1
#> 7 independent multinomial              0           1  15.1
#> 8 independent multinomial              0           1  15.1
#>   method                              log_e_bf10 expression
#>   <chr>                                    <dbl> <list>    
#> 1 Bayesian contingency table analysis       2.72 <language>
#> 2 Bayesian contingency table analysis       2.72 <language>
#> 3 Bayesian contingency table analysis       2.72 <language>
#> 4 Bayesian contingency table analysis       2.72 <language>
#> 5 Bayesian contingency table analysis       2.72 <language>
#> 6 Bayesian contingency table analysis       2.72 <language>
#> 7 Bayesian contingency table analysis       2.72 <language>
#> 8 Bayesian contingency table analysis       2.72 <language>
#> 
#> $pairwise_comparisons_data
#> NULL
#> 
#> $descriptive_data
#> # A tibble: 6 x 5
#>   am    cyl   counts  perc .label
#>   <fct> <fct>  <int> <dbl> <chr> 
#> 1 0     8         12  63.2 63%   
#> 2 1     8          2  15.4 15%   
#> 3 0     6          4  21.1 21%   
#> 4 1     6          3  23.1 23%   
#> 5 0     4          3  15.8 16%   
#> 6 1     4          8  61.5 62%   
#> 
#> $one_sample_data
#> # A tibble: 2 x 10
#>   am    counts  perc N        statistic    df p.value
#>   <fct>  <int> <dbl> <chr>        <dbl> <dbl>   <dbl>
#> 1 1         13  40.6 (n = 13)      4.77     2  0.0921
#> 2 0         19  59.4 (n = 19)      7.68     2  0.0214
#>   method                                  
#>   <chr>                                   
#> 1 Chi-squared test for given probabilities
#> 2 Chi-squared test for given probabilities
#>   .label                                                            
#>   <chr>                                                             
#> 1 list(~chi['gof']^2~(2)==4.77, ~italic(p)=='0.092', ~italic(n)==13)
#> 2 list(~chi['gof']^2~(2)==7.68, ~italic(p)=='0.021', ~italic(n)==19)
#>   .p.label                 
#>   <chr>                    
#> 1 list(~italic(p)=='0.092')
#> 2 list(~italic(p)=='0.021')
#> 
# }