Extracting dataframes with statistical details from {ggstatsplot}

Extracting dataframes with statistical details from {ggstatsplot}

Usage

extract_stats(p, ...)

Arguments

p

A plot from {ggstatsplot} package

...

Ignored

Value

A list of tibbles containing statistical analysis summaries.

Details

This is a convenience function to extract dataframes with statistical details that are used to create expressions displayed in {ggstatsplot} plots as subtitle and/or as caption. Note that all of this analysis is carried out by the {statsExpressions} package.

For more details about underlying tests and effect size estimates, see the following vignette: https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html

Examples

# \donttest{
set.seed(123)
library(ggstatsplot)

# in case of group comparisons
p <- ggbetweenstats(mtcars, cyl, mpg)
extract_stats(p)
#> $subtitle_data
#> # A tibble: 1 × 13
#>   statistic    df df.error    p.value method      effectsize estimate conf.level
#>       <dbl> <dbl>    <dbl>      <dbl> <chr>       <chr>         <dbl>      <dbl>
#> 1      31.6     2     18.0 0.00000127 One-way an… Omega2        0.744       0.95
#> # … with 5 more variables: conf.low <dbl>, conf.high <dbl>, conf.method <chr>,
#> #   conf.distribution <chr>, expression <list>
#> 
#> $caption_data
#> # A tibble: 6 × 20
#>   term  estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>    <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 mu      20.5         0.95  19.3        21.9  1              0     
#> 2 cyl-4    5.90        0.95   4.11        7.52 1              0     
#> 3 cyl-6   -0.704       0.95  -2.64        1.06 0.780          0.390 
#> 4 cyl-8   -5.18        0.95  -6.76       -3.55 1              0     
#> 5 sig2    11.0         0.95   6.24       18.3  1              0     
#> 6 g_cyl    2.69        0.95   0.0911     18.7  1              0.0418
#> # … with 13 more variables: prior.distribution <chr>, prior.location <dbl>,
#> #   prior.scale <dbl>, component <chr>, bf10 <dbl>, method <chr>,
#> #   log_e_bf10 <dbl>, r2 <dbl>, std.dev <dbl>, r2.conf.level <dbl>,
#> #   r2.conf.low <dbl>, r2.conf.high <dbl>, expression <list>
#> 
#> $pairwise_comparisons_data
#> # A tibble: 3 × 11
#>   group1 group2 statistic   p.value alternative method distribution p.adjustment
#>   <chr>  <chr>      <dbl>     <dbl> <chr>       <chr>  <chr>        <chr>       
#> 1 4      6          -6.67 0.00110   two.sided   Games… q            none        
#> 2 4      8         -10.7  0.0000140 two.sided   Games… q            none        
#> 3 6      8          -7.48 0.000257  two.sided   Games… q            none        
#> # … with 3 more variables: test.details <chr>, p.value.adjustment <chr>,
#> #   label <chr>
#> 
#> $descriptive_data
#> NULL
#> 
#> $one_sample_data
#> NULL
#> 

# the exact details depend on the function
extract_stats(ggbarstats(mtcars, cyl, am))
#> $subtitle_data
#> # A tibble: 1 × 12
#>   statistic    df p.value method       effectsize   estimate conf.level conf.low
#>       <dbl> <int>   <dbl> <chr>        <chr>           <dbl>      <dbl>    <dbl>
#> 1      8.74     2  0.0126 Pearson's C… Cramer's V …    0.464       0.95        0
#> # … with 4 more variables: conf.high <dbl>, conf.method <chr>,
#> #   conf.distribution <chr>, expression <list>
#> 
#> $caption_data
#> NULL
#> 
#> $pairwise_comparisons_data
#> NULL
#> 
#> $descriptive_data
#> # A tibble: 6 × 5
#>   am    cyl   counts  perc .label
#>   <fct> <fct>  <int> <dbl> <chr> 
#> 1 0     8         12  63.2 63%   
#> 2 1     8          2  15.4 15%   
#> 3 0     6          4  21.1 21%   
#> 4 1     6          3  23.1 23%   
#> 5 0     4          3  15.8 16%   
#> 6 1     4          8  61.5 62%   
#> 
#> $one_sample_data
#> # A tibble: 2 × 10
#>   am    counts  perc N        statistic    df p.value method   .label   .p.label
#>   <fct>  <int> <dbl> <chr>        <dbl> <dbl>   <dbl> <chr>    <chr>    <chr>   
#> 1 1         13  40.6 (n = 13)      4.77     2  0.0921 Chi-squ… list(~c… list(~i…
#> 2 0         19  59.4 (n = 19)      7.68     2  0.0214 Chi-squ… list(~c… list(~i…
#> 
# }