Extracting data frames or expressions from {ggstatsplot} plots

Extracting data frames or expressions from {ggstatsplot} plots

Usage

extract_stats(p, ...)

extract_subtitle(p)

extract_caption(p)

Arguments

p

A plot from {ggstatsplot} package

...

Ignored

Value

A list of tibbles containing summaries of various statistical analyses. The exact details included will depend on the function.

Details

These are convenience functions to extract data frames or expressions with statistical details that are used to create expressions displayed in {ggstatsplot} plots as subtitle, caption, etc. Note that all of this analysis is carried out by the {statsExpressions} package. And so if you are using these functions only to extract data frames, you are better off using that package.

The only exception is the ggcorrmat() function. But, if a data frame is what you want, you shouldn't be using ggcorrmat() anyway. You can use correlation::correlation() function which provides tidy data frames by default.

Examples

# \donttest{
set.seed(123)
library(ggstatsplot)

# non-grouped function -----------------------------

# in case of group comparisons
p <- ggbetweenstats(mtcars, cyl, mpg)
extract_stats(p)
#> $subtitle_data
#> # A tibble: 1 × 14
#>   statistic    df df.er…¹ p.value method effec…² estim…³ conf.…⁴ conf.…⁵ conf.…⁶
#>       <dbl> <dbl>   <dbl>   <dbl> <chr>  <chr>     <dbl>   <dbl>   <dbl>   <dbl>
#> 1      31.6     2    18.0 1.27e-6 One-w… Omega2    0.744    0.95   0.531       1
#> # … with 4 more variables: conf.method <chr>, conf.distribution <chr>,
#> #   n.obs <int>, expression <list>, and abbreviated variable names ¹​df.error,
#> #   ²​effectsize, ³​estimate, ⁴​conf.level, ⁵​conf.low, ⁶​conf.high
#> 
#> $caption_data
#> # A tibble: 6 × 17
#>   term     pd prior.dist…¹ prior…² prior…³   bf10 method log_e…⁴ effec…⁵ estim…⁶
#>   <chr> <dbl> <chr>          <dbl>   <dbl>  <dbl> <chr>    <dbl> <chr>     <dbl>
#> 1 mu    1     cauchy             0   0.707 3.01e6 Bayes…    14.9 Bayesi…   0.714
#> 2 cyl-4 1     cauchy             0   0.707 3.01e6 Bayes…    14.9 Bayesi…   0.714
#> 3 cyl-6 0.780 cauchy             0   0.707 3.01e6 Bayes…    14.9 Bayesi…   0.714
#> 4 cyl-8 1     cauchy             0   0.707 3.01e6 Bayes…    14.9 Bayesi…   0.714
#> 5 sig2  1     cauchy             0   0.707 3.01e6 Bayes…    14.9 Bayesi…   0.714
#> 6 g_cyl 1     cauchy             0   0.707 3.01e6 Bayes…    14.9 Bayesi…   0.714
#> # … with 7 more variables: std.dev <dbl>, conf.level <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, conf.method <chr>, n.obs <int>, expression <list>, and
#> #   abbreviated variable names ¹​prior.distribution, ²​prior.location,
#> #   ³​prior.scale, ⁴​log_e_bf10, ⁵​effectsize, ⁶​estimate
#> 
#> $pairwise_comparisons_data
#> # A tibble: 3 × 9
#>   group1 group2 statistic   p.value alternative distr…¹ p.adj…² test  expression
#>   <chr>  <chr>      <dbl>     <dbl> <chr>       <chr>   <chr>   <chr> <list>    
#> 1 4      6          -6.67 0.00110   two.sided   q       Holm    Game… <language>
#> 2 4      8         -10.7  0.0000140 two.sided   q       Holm    Game… <language>
#> 3 6      8          -7.48 0.000257  two.sided   q       Holm    Game… <language>
#> # … with abbreviated variable names ¹​distribution, ²​p.adjust.method
#> 
#> $descriptive_data
#> NULL
#> 
#> $one_sample_data
#> NULL
#> 
#> $tidy_data
#> NULL
#> 
#> $glance_data
#> NULL
#> 

# the exact details depend on the function
extract_stats(ggbarstats(mtcars, cyl, am))
#> $subtitle_data
#> # A tibble: 1 × 13
#>   statistic    df p.value method effec…¹ estim…² conf.…³ conf.…⁴ conf.…⁵ conf.…⁶
#>       <dbl> <int>   <dbl> <chr>  <chr>     <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#> 1      8.74     2  0.0126 Pears… Cramer…   0.464    0.95       0       1 ncp    
#> # … with 3 more variables: conf.distribution <chr>, n.obs <int>,
#> #   expression <list>, and abbreviated variable names ¹​effectsize, ²​estimate,
#> #   ³​conf.level, ⁴​conf.low, ⁵​conf.high, ⁶​conf.method
#> 
#> $caption_data
#> # A tibble: 1 × 15
#>   term  conf.level effec…¹ estim…² conf.…³ conf.…⁴ prior…⁵ prior…⁶ prior…⁷  bf10
#>   <chr>      <dbl> <chr>     <dbl>   <dbl>   <dbl> <chr>     <dbl>   <dbl> <dbl>
#> 1 Ratio       0.95 Cramer…   0.410       0   0.670 indepe…       0       1  15.1
#> # … with 5 more variables: method <chr>, conf.method <chr>, log_e_bf10 <dbl>,
#> #   n.obs <int>, expression <list>, and abbreviated variable names ¹​effectsize,
#> #   ²​estimate, ³​conf.low, ⁴​conf.high, ⁵​prior.distribution, ⁶​prior.location,
#> #   ⁷​prior.scale
#> 
#> $pairwise_comparisons_data
#> NULL
#> 
#> $descriptive_data
#> # A tibble: 6 × 5
#>   am    cyl   counts  perc .label
#>   <fct> <fct>  <int> <dbl> <chr> 
#> 1 0     8         12  63.2 63%   
#> 2 1     8          2  15.4 15%   
#> 3 0     6          4  21.1 21%   
#> 4 1     6          3  23.1 23%   
#> 5 0     4          3  15.8 16%   
#> 6 1     4          8  61.5 62%   
#> 
#> $one_sample_data
#> # A tibble: 2 × 10
#>   am    counts  perc N        statistic    df p.value method      .label .p.la…¹
#>   <fct>  <int> <dbl> <chr>        <dbl> <dbl>   <dbl> <chr>       <chr>  <chr>  
#> 1 1         13  40.6 (n = 13)      4.77     2  0.0921 Chi-square… list(… list(~…
#> 2 0         19  59.4 (n = 19)      7.68     2  0.0214 Chi-square… list(… list(~…
#> # … with abbreviated variable name ¹​.p.label
#> 
#> $tidy_data
#> NULL
#> 
#> $glance_data
#> NULL
#> 

# grouped function -----------------------------
p <- grouped_ggbarstats(
  Titanic_full,
  x = Survived,
  y = Sex,
  grouping.var = Age
)

extract_stats(p[[1L]])
#> $subtitle_data
#> # A tibble: 1 × 13
#>   stati…¹    df   p.value method effec…² estim…³ conf.…⁴ conf.…⁵ conf.…⁶ conf.…⁷
#>     <dbl> <int>     <dbl> <chr>  <chr>     <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#> 1    461.     1 3.11e-102 Pears… Cramer…   0.469    0.95   0.433       1 ncp    
#> # … with 3 more variables: conf.distribution <chr>, n.obs <int>,
#> #   expression <list>, and abbreviated variable names ¹​statistic, ²​effectsize,
#> #   ³​estimate, ⁴​conf.level, ⁵​conf.low, ⁶​conf.high, ⁷​conf.method
#> 
#> $caption_data
#> # A tibble: 1 × 15
#>   term  conf.l…¹ effec…² estim…³ conf.…⁴ conf.…⁵ prior…⁶ prior…⁷ prior…⁸    bf10
#>   <chr>    <dbl> <chr>     <dbl>   <dbl>   <dbl> <chr>     <dbl>   <dbl>   <dbl>
#> 1 Ratio     0.95 Cramer…   0.468   0.427   0.510 indepe…       0       1 7.02e92
#> # … with 5 more variables: method <chr>, conf.method <chr>, log_e_bf10 <dbl>,
#> #   n.obs <int>, expression <list>, and abbreviated variable names ¹​conf.level,
#> #   ²​effectsize, ³​estimate, ⁴​conf.low, ⁵​conf.high, ⁶​prior.distribution,
#> #   ⁷​prior.location, ⁸​prior.scale
#> 
#> $pairwise_comparisons_data
#> NULL
#> 
#> $descriptive_data
#> # A tibble: 4 × 5
#>   Sex    Survived counts  perc .label
#>   <fct>  <fct>     <int> <dbl> <chr> 
#> 1 Female Yes         316  74.4 74%   
#> 2 Male   Yes         338  20.3 20%   
#> 3 Female No          109  25.6 26%   
#> 4 Male   No         1329  79.7 80%   
#> 
#> $one_sample_data
#> # A tibble: 2 × 10
#>   Sex    counts  perc N           statis…¹    df   p.value method .label .p.la…²
#>   <fct>   <int> <dbl> <chr>          <dbl> <dbl>     <dbl> <chr>  <chr>  <chr>  
#> 1 Male     1667  79.7 (n = 1,667)     589.     1 3.87e-130 Chi-s… list(… list(~…
#> 2 Female    425  20.3 (n = 425)       101.     1 1.01e- 23 Chi-s… list(… list(~…
#> # … with abbreviated variable names ¹​statistic, ²​.p.label
#> 
#> $tidy_data
#> NULL
#> 
#> $glance_data
#> NULL
#> 
extract_stats(p[[2L]])
#> $subtitle_data
#> # A tibble: 1 × 13
#>   statistic    df p.value method effec…¹ estim…² conf.…³ conf.…⁴ conf.…⁵ conf.…⁶
#>       <dbl> <int>   <dbl> <chr>  <chr>     <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#> 1      3.03     1  0.0818 Pears… Cramer…   0.137    0.95       0       1 ncp    
#> # … with 3 more variables: conf.distribution <chr>, n.obs <int>,
#> #   expression <list>, and abbreviated variable names ¹​effectsize, ²​estimate,
#> #   ³​conf.level, ⁴​conf.low, ⁵​conf.high, ⁶​conf.method
#> 
#> $caption_data
#> # A tibble: 1 × 15
#>   term  conf.level effec…¹ estim…² conf.…³ conf.…⁴ prior…⁵ prior…⁶ prior…⁷  bf10
#>   <chr>      <dbl> <chr>     <dbl>   <dbl>   <dbl> <chr>     <dbl>   <dbl> <dbl>
#> 1 Ratio       0.95 Cramer…   0.129       0   0.330 indepe…       0       1  1.03
#> # … with 5 more variables: method <chr>, conf.method <chr>, log_e_bf10 <dbl>,
#> #   n.obs <int>, expression <list>, and abbreviated variable names ¹​effectsize,
#> #   ²​estimate, ³​conf.low, ⁴​conf.high, ⁵​prior.distribution, ⁶​prior.location,
#> #   ⁷​prior.scale
#> 
#> $pairwise_comparisons_data
#> NULL
#> 
#> $descriptive_data
#> # A tibble: 4 × 5
#>   Sex    Survived counts  perc .label
#>   <fct>  <fct>     <int> <dbl> <chr> 
#> 1 Female Yes          28  62.2 62%   
#> 2 Male   Yes          29  45.3 45%   
#> 3 Female No           17  37.8 38%   
#> 4 Male   No           35  54.7 55%   
#> 
#> $one_sample_data
#> # A tibble: 2 × 10
#>   Sex    counts  perc N        statistic    df p.value method     .label .p.la…¹
#>   <fct>   <int> <dbl> <chr>        <dbl> <dbl>   <dbl> <chr>      <chr>  <chr>  
#> 1 Male       64  58.7 (n = 64)     0.562     1   0.453 Chi-squar… list(… list(~…
#> 2 Female     45  41.3 (n = 45)     2.69      1   0.101 Chi-squar… list(… list(~…
#> # … with abbreviated variable name ¹​.p.label
#> 
#> $tidy_data
#> NULL
#> 
#> $glance_data
#> NULL
#> 
# }