Create labels with statistical details for ggcoefstats.

Create labels with statistical details for ggcoefstats.

ggcoefstats_label_maker(x, tidy_df = NULL, glance_df = NULL,
  statistic = NULL, k = 2, effsize = "eta", partial = TRUE)

Arguments

x
A model object to be tidied with broom::tidy, or a tidy data frame containing results. If a data frame is to be plotted, it must contain columns named term (names of predictors), or estimate (corresponding estimates of coefficients or other quantities of interest). Other optional columns are conf.low and conf.high (for confidence intervals); p.value.

tidy_df
Tidy dataframe from broom::tidy.

glance_df
Glance model summary dataframe from broom::glance (default: NULL). This is optional argument. If provide, the glance summary will be used to write caption for the final plot.

statistic
Which statistic is to be displayed (either "t" or "f"or "z") in the label. This is especially important if the x argument in ggcoefstats is a dataframe in which case the function wouldn't know what kind of model it is dealing with.

k
Number of decimal places expected for results displayed in labels (Default : k = 2).

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega". This argument is relevant only for models objects of class aov, anova, and aovlist.

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class aov, anova, or aovlist.

Examples

# show all columns in a tibble
options(tibble.width = Inf)

# for reproducibility
set.seed(123)

#------------------------- models with *t*-statistic ------------------
# model with t-statistic
ggstatsplot:::ggcoefstats_label_maker(x = broom::tidy(stats::lm(
  data = mtcars, formula = wt ~ cyl * mpg
)), statistic = "t")

#>          term     estimate  std.error statistic    p.value significance
#> 1 (Intercept)  2.982425775 1.57304118      1.90 0.06832893           ns
#> 2         cyl  0.477603816 0.23345602      2.05 0.05027309           ns
#> 3         mpg -0.009473015 0.06598555     -0.14 0.88687461           ns
#> 4     cyl:mpg -0.021919260 0.01202435     -1.82 0.07901275           ns
#>   p.value.formatted p.value.formatted2
#> 1             0.068            ==0.068
#> 2             0.050            ==0.050
#> 3             0.887            ==0.887
#> 4             0.079            ==0.079
#>                                                              label
#> 1   list(~italic(beta)==2.98, ~italic(t)==1.90, ~italic(p)==0.068)
#> 2   list(~italic(beta)==0.48, ~italic(t)==2.05, ~italic(p)==0.050)
#> 3 list(~italic(beta)==-0.01, ~italic(t)==-0.14, ~italic(p)==0.887)
#> 4 list(~italic(beta)==-0.02, ~italic(t)==-1.82, ~italic(p)==0.079)

# (in case `x` is not a dataframe, no need to specify `statistic` argument;
# this will be figured out by the function itself)

#------------------------- models with *t*-statistic ------------------

# dataframe
clotting <- data.frame(
  u = c(5, 10, 15, 20, 30, 40, 60, 80, 100),
  lot1 = c(118, 58, 42, 35, 27, 25, 21, 19, 18),
  lot2 = c(69, 35, 26, 21, 18, 16, 13, 12, 12)
)

# model
mod <-
  stats::glm(
    formula = lot1 ~ log(u),
    data = clotting,
    family = Gamma
  )

# model with t-statistic
ggstatsplot:::ggcoefstats_label_maker(
  x = mod,
  tidy_df = broom::tidy(
    x = mod,
    conf.int = TRUE,
    conf.level = 0.95
  )
)

#>          term    estimate    std.error statistic    conf.low   conf.high
#> 1 (Intercept) -0.01655438 0.0009275466    -17.85 -0.01835528 -0.01471815
#> 2      log(u)  0.01534311 0.0004149596     36.97  0.01453306  0.01615996
#>        p.value significance p.value.formatted p.value.formatted2
#> 1 4.279149e-07          ***           < 0.001           <= 0.001
#> 2 2.751191e-09          ***           < 0.001           <= 0.001
#>                                                                label
#> 1 list(~italic(beta)==-0.02, ~italic(t)==-17.85, ~italic(p)<= 0.001)
#> 2   list(~italic(beta)==0.02, ~italic(t)==36.97, ~italic(p)<= 0.001)

#------------------------- models with *z*-statistic --------------------

# preparing dataframe
counts <- c(18, 17, 15, 20, 10, 20, 25, 13, 12)
outcome <- gl(3, 1, 9)
treatment <- gl(3, 3)
d.AD <- data.frame(treatment, outcome, counts)

# model
mod <- stats::glm(
  formula = counts ~ outcome + treatment,
  family = poisson(),
  data = d.AD
)

# creating tidy dataframe with label column
ggstatsplot:::ggcoefstats_label_maker(x = mod, tidy_df = broom::tidy(mod))

#>          term      estimate std.error statistic      p.value significance
#> 1 (Intercept)  3.044522e+00 0.1708987     17.81 5.426767e-71          ***
#> 2    outcome2 -4.542553e-01 0.2021708     -2.25 2.464711e-02            *
#> 3    outcome3 -2.929871e-01 0.1927423     -1.52 1.284865e-01           ns
#> 4  treatment2  1.337909e-15 0.2000000      0.00 1.000000e+00           ns
#> 5  treatment3  1.421085e-15 0.2000000      0.00 1.000000e+00           ns
#>   p.value.formatted p.value.formatted2
#> 1           < 0.001           <= 0.001
#> 2             0.025            ==0.025
#> 3             0.128            ==0.128
#> 4             1.000            ==1.000
#> 5             1.000            ==1.000
#>                                                              label
#> 1 list(~italic(beta)==3.04, ~italic(z)==17.81, ~italic(p)<= 0.001)
#> 2 list(~italic(beta)==-0.45, ~italic(z)==-2.25, ~italic(p)==0.025)
#> 3 list(~italic(beta)==-0.29, ~italic(z)==-1.52, ~italic(p)==0.128)
#> 4   list(~italic(beta)==0.00, ~italic(z)==0.00, ~italic(p)==1.000)
#> 5   list(~italic(beta)==0.00, ~italic(z)==0.00, ~italic(p)==1.000)

#------------------------- models with *f*-statistic --------------------
# creating a model object
op <- options(contrasts = c("contr.helmert", "contr.poly"))
npk.aov <- stats::aov(formula = yield ~ block + N * P * K, data = npk)

# converting to a dataframe using
tidy_df <- ggstatsplot::lm_effsize_ci(
  object = npk.aov,
  effsize = "omega",
  partial = FALSE,
  nboot = 50
) %>%
  dplyr::rename(.data = ., estimate = omegasq, statistic = F.value)

# including a new column with a label
ggstatsplot:::ggcoefstats_label_maker(
  x = npk.aov,
  tidy_df = tidy_df,
  effsize = "omega",
  partial = FALSE
)

#>    term statistic df1 df2     estimate   conf.low conf.high     p.value
#> 1 block      4.45   5  12  0.298374708 0.02983123 0.6749877 0.015938790
#> 2     N     12.26   1  12  0.194931631 0.04403242 0.5880499 0.004371812
#> 3     P      0.54   1  12 -0.007892852         NA 0.2327486 0.474904093
#> 4     K      6.17   1  12  0.089437782         NA 0.4722947 0.028795054
#> 5   N:P      1.38   1  12  0.006549759         NA 0.2944957 0.263165283
#> 6   N:K      2.15   1  12  0.019841146         NA 0.3355048 0.168647879
#> 7   P:K      0.03   1  12 -0.016773711         NA 0.1255008 0.862752086
#>   significance p.value.formatted p.value.formatted2
#> 1            *             0.016            ==0.016
#> 2           **             0.004            ==0.004
#> 3           ns             0.475            ==0.475
#> 4            *             0.029            ==0.029
#> 5           ns             0.263            ==0.263
#> 6           ns             0.169            ==0.169
#> 7           ns             0.863            ==0.863
#>                                                                          label
#> 1  list(~italic(F)(5*","*12)==4.45, ~italic(p)==0.016, ~italic(omega)^2==0.30)
#> 2 list(~italic(F)(1*","*12)==12.26, ~italic(p)==0.004, ~italic(omega)^2==0.19)
#> 3 list(~italic(F)(1*","*12)==0.54, ~italic(p)==0.475, ~italic(omega)^2==-0.01)
#> 4  list(~italic(F)(1*","*12)==6.17, ~italic(p)==0.029, ~italic(omega)^2==0.09)
#> 5  list(~italic(F)(1*","*12)==1.38, ~italic(p)==0.263, ~italic(omega)^2==0.01)
#> 6  list(~italic(F)(1*","*12)==2.15, ~italic(p)==0.169, ~italic(omega)^2==0.02)
#> 7 list(~italic(F)(1*","*12)==0.03, ~italic(p)==0.863, ~italic(omega)^2==-0.02)