Create labels with statistical details for ggcoefstats.

Create labels with statistical details for ggcoefstats.

ggcoefstats_label_maker(x, tidy_df = NULL, glance_df = NULL,
  statistic = NULL, k = 2, effsize = "eta", partial = TRUE)

Arguments

x
A model object to be tidied with broom::tidy, or a tidy data frame containing results. If a data frame is to be plotted, it must contain columns named term (names of predictors), or estimate (corresponding estimates of coefficients or other quantities of interest). Other optional columns are conf.low and conf.high (for confidence intervals); p.value. It is important that all term names should be unique.

tidy_df
Tidy dataframe from broomExtra::tidy.

glance_df
Glance model summary dataframe from broom::glance (default: NULL). This is optional argument. If provide, the glance summary will be used to write caption for the final plot.

statistic
Which statistic is to be displayed (either "t" or "f"or "z") in the label. This is especially important if the x argument in ggcoefstats is a dataframe in which case the function wouldn't know what kind of model it is dealing with.

k
Number of decimal places expected for results displayed in labels (Default : k = 2).

effsize
Character describing the effect size to be displayed: "eta" (default) or "omega". This argument is relevant only for models objects of class aov, anova, and aovlist.

partial
Logical that decides if partial eta-squared or omega-squared are returned (Default: TRUE). If FALSE, eta-squared or omega-squared will be returned. Valid only for objects of class aov, anova, or aovlist.

Examples

# show all columns in output tibble
options(tibble.width = Inf)

# for reproducibility
set.seed(123)

#------------------------- models with *t*-statistic ------------------
# model with t-statistic
ggstatsplot:::ggcoefstats_label_maker(x = broomExtra::tidy(stats::lm(
  data = mtcars, formula = wt ~ cyl * mpg
)), statistic = "t")

#> # A tibble: 4 x 9
#>   term        estimate std.error statistic p.value significance
#>   <chr>          <dbl>     <dbl> <chr>       <dbl> <chr>       
#> 1 (Intercept)  2.98       1.57   1.90       0.0683 ns          
#> 2 cyl          0.478      0.233  2.05       0.0503 ns          
#> 3 mpg         -0.00947    0.0660 -0.14      0.887  ns          
#> 4 cyl:mpg     -0.0219     0.0120 -1.82      0.0790 ns          
#>   p.value.formatted p.value.formatted2
#>   <chr>             <chr>             
#> 1 0.068             ==0.068           
#> 2 0.050             ==0.050           
#> 3 0.887             ==0.887           
#> 4 0.079             ==0.079           
#>   label                                                           
#>   <chr>                                                           
#> 1 list(~italic(beta)==2.98, ~italic(t)==1.90, ~italic(p)==0.068)  
#> 2 list(~italic(beta)==0.48, ~italic(t)==2.05, ~italic(p)==0.050)  
#> 3 list(~italic(beta)==-0.01, ~italic(t)==-0.14, ~italic(p)==0.887)
#> 4 list(~italic(beta)==-0.02, ~italic(t)==-1.82, ~italic(p)==0.079)

# (in case `x` is not a dataframe, no need to specify `statistic` argument;
# this will be figured out by the function itself)

#------------------------- models with *t*-statistic ------------------

# dataframe
clotting <- data.frame(
  u = c(5, 10, 15, 20, 30, 40, 60, 80, 100),
  lot1 = c(118, 58, 42, 35, 27, 25, 21, 19, 18),
  lot2 = c(69, 35, 26, 21, 18, 16, 13, 12, 12)
)

# model
mod <-
  stats::glm(
    formula = lot1 ~ log(u),
    data = clotting,
    family = Gamma
  )

# model with t-statistic
ggstatsplot:::ggcoefstats_label_maker(
  x = mod,
  tidy_df = broomExtra::tidy(
    x = mod,
    conf.int = TRUE,
    conf.level = 0.95
  )
)

#> # A tibble: 2 x 11
#>   term        estimate std.error statistic conf.low conf.high       p.value
#>   <chr>          <dbl>     <dbl> <chr>        <dbl>     <dbl>         <dbl>
#> 1 (Intercept)  -0.0166  0.000928 -17.85     -0.0184   -0.0147 0.000000428  
#> 2 log(u)        0.0153  0.000415 36.97       0.0145    0.0162 0.00000000275
#>   significance p.value.formatted p.value.formatted2
#>   <chr>        <chr>             <chr>             
#> 1 ***          < 0.001           <= 0.001          
#> 2 ***          < 0.001           <= 0.001          
#>   label                                                             
#>   <chr>                                                             
#> 1 list(~italic(beta)==-0.02, ~italic(t)==-17.85, ~italic(p)<= 0.001)
#> 2 list(~italic(beta)==0.02, ~italic(t)==36.97, ~italic(p)<= 0.001)  

#------------------------- models with *z*-statistic --------------------

# preparing dataframe
counts <- c(18, 17, 15, 20, 10, 20, 25, 13, 12)
outcome <- gl(3, 1, 9)
treatment <- gl(3, 3)
d.AD <- data.frame(treatment, outcome, counts)

# model
mod <- stats::glm(
  formula = counts ~ outcome + treatment,
  family = poisson(),
  data = d.AD
)

# creating tidy dataframe with label column
ggstatsplot:::ggcoefstats_label_maker(x = mod, tidy_df = broomExtra::tidy(mod))

#> # A tibble: 5 x 9
#>   term         estimate std.error statistic   p.value significance
#>   <chr>           <dbl>     <dbl> <chr>         <dbl> <chr>       
#> 1 (Intercept)  3.04e+ 0     0.171 17.81      5.43e-71 ***         
#> 2 outcome2    -4.54e- 1     0.202 -2.25      2.46e- 2 *           
#> 3 outcome3    -2.93e- 1     0.193 -1.52      1.28e- 1 ns          
#> 4 treatment2   1.34e-15     0.200 0.00      10.00e- 1 ns          
#> 5 treatment3   1.42e-15     0.200 0.00      10.00e- 1 ns          
#>   p.value.formatted p.value.formatted2
#>   <chr>             <chr>             
#> 1 < 0.001           <= 0.001          
#> 2 0.025             ==0.025           
#> 3 0.128             ==0.128           
#> 4 1.000             ==1.000           
#> 5 1.000             ==1.000           
#>   label                                                           
#>   <chr>                                                           
#> 1 list(~italic(beta)==3.04, ~italic(z)==17.81, ~italic(p)<= 0.001)
#> 2 list(~italic(beta)==-0.45, ~italic(z)==-2.25, ~italic(p)==0.025)
#> 3 list(~italic(beta)==-0.29, ~italic(z)==-1.52, ~italic(p)==0.128)
#> 4 list(~italic(beta)==0.00, ~italic(z)==0.00, ~italic(p)==1.000)  
#> 5 list(~italic(beta)==0.00, ~italic(z)==0.00, ~italic(p)==1.000)  

#------------------------- models with *f*-statistic --------------------
# creating a model object
op <- options(contrasts = c("contr.helmert", "contr.poly"))
npk.aov <- stats::aov(formula = yield ~ block + N * P * K, data = npk)

# extracting a tidy dataframe with effect size estimate and their CIs
tidy_df <-
  ggstatsplot::lm_effsize_ci(
    object = npk.aov,
    effsize = "omega",
    partial = FALSE,
    nboot = 50
  ) %>%
  dplyr::rename(.data = ., estimate = omegasq, statistic = F.value)

# including a new column with a label
ggstatsplot:::ggcoefstats_label_maker(
  x = npk.aov,
  tidy_df = tidy_df,
  effsize = "omega",
  partial = FALSE
)

#> # A tibble: 7 x 13
#>   term  statistic   df1   df2 estimate conf.low conf.high p.value significance
#>   <chr> <chr>     <dbl> <dbl>    <dbl>    <dbl>     <dbl>   <dbl> <chr>       
#> 1 block 4.45          5    12  0.298     0.0298     0.675 0.0159  *           
#> 2 N     12.26         1    12  0.195     0.0440     0.588 0.00437 **          
#> 3 P     0.54          1    12 -0.00789  NA          0.233 0.475   ns          
#> 4 K     6.17          1    12  0.0894   NA          0.472 0.0288  *           
#> 5 N:P   1.38          1    12  0.00655  NA          0.294 0.263   ns          
#> 6 N:K   2.15          1    12  0.0198   NA          0.336 0.169   ns          
#> 7 P:K   0.03          1    12 -0.0168   NA          0.126 0.863   ns          
#>   p.value.formatted p.value.formatted2 effsize.text
#>   <chr>             <chr>              <list>      
#> 1 0.016             ==0.016            <language>  
#> 2 0.004             ==0.004            <language>  
#> 3 0.475             ==0.475            <language>  
#> 4 0.029             ==0.029            <language>  
#> 5 0.263             ==0.263            <language>  
#> 6 0.169             ==0.169            <language>  
#> 7 0.863             ==0.863            <language>  
#>   label                                                                         
#>   <chr>                                                                         
#> 1 "list(~italic(F)(5*\",\"*12)==4.45, ~italic(p)==0.016, ~italic(omega)^2==0.30~
#> 2 "list(~italic(F)(1*\",\"*12)==12.26, ~italic(p)==0.004, ~italic(omega)^2==0.1~
#> 3 "list(~italic(F)(1*\",\"*12)==0.54, ~italic(p)==0.475, ~italic(omega)^2==-0.0~
#> 4 "list(~italic(F)(1*\",\"*12)==6.17, ~italic(p)==0.029, ~italic(omega)^2==0.09~
#> 5 "list(~italic(F)(1*\",\"*12)==1.38, ~italic(p)==0.263, ~italic(omega)^2==0.01~
#> 6 "list(~italic(F)(1*\",\"*12)==2.15, ~italic(p)==0.169, ~italic(omega)^2==0.02~
#> 7 "list(~italic(F)(1*\",\"*12)==0.03, ~italic(p)==0.863, ~italic(omega)^2==-0.0~