ggstatsplot: 'ggplot2' Based Plots with Statistical Details

ggstatsplot is an extension of ggplot2 package for creating graphics with details from statistical tests included in the plots themselves and targeted primarily at behavioral sciences community to provide a one-line code to produce information-rich plots. In a typical exploratory data analysis workflow, data visualization and statistical modeling are two different phases: visualization informs modeling, and modeling in its turn can suggest a different visualization method, and so on and so forth. The central idea of ggstatsplot is simple: combine these two phases into one in the form of graphics with statistical details, which makes data exploration simpler and faster. Currently, it supports only the most common types of statistical tests (parametric, non-parametric, Bayes Factor, and robust versions of t-test/anova, correlation, regression analyses, contingency tables analyses, and meta-analysis).

Details

ggstatsplot

The main functions are-

ggbetweenstats function to produce information-rich comparison plot between different groups or conditions with ggplot2 and details from the statistical tests in the subtitle.

ggwithinstats function to produce information-rich comparison plot within different groups or conditions with ggplot2 and details from the statistical tests in the subtitle.

ggscatterstats function to produce ggplot2 scatterplots along with a marginal histograms/boxplots/density plots from ggExtra and details from the statistical tests in the subtitle.

ggpiestats function to produce pie chart with details from the statistical tests in the subtitle.

ggbarstats function to produce stacked bar chart with details from the statistical tests in the subtitle.

gghistostats function to produce histogram for a single variable with results from one sample test displayed in the subtitle.

ggdotplotstats function to produce Cleveland-style dot plots/charts for a single variable with labels and results from one sample test displayed in the subtitle.

ggcorrmat function to visualize the correlation matrix.

ggcoefstats function to visualize results from regression analyses.

combine_plots2 helper function to combine multiple ggstatsplot plots using cowplot::plot_grid() with a combination of title, caption, and annotation label.

For more documentation, see the dedicated Website.

See also