Split dataframe into a list by grouping variable.

This function splits the dataframe into a list, with the length of the list equal to the factor levels of the grouping variable. Each element of the list will be a tibble.

Usage

grouped_list(data, grouping.var = NULL)

Arguments

data

A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

grouping.var

A single grouping variable.

Examples

# \donttest{
ggstatsplot:::grouped_list(ggplot2::msleep, grouping.var = vore)
#> $carni
#> # A tibble: 19 x 11
#>    name                       genus         vore  order           conservation sleep_total sleep_rem sleep_cycle awake
#>    <chr>                      <chr>         <chr> <chr>           <chr>              <dbl>     <dbl>       <dbl> <dbl>
#>  1 Cheetah                    Acinonyx      carni Carnivora       lc                  12.1      NA        NA      11.9
#>  2 Northern fur seal          Callorhinus   carni Carnivora       vu                   8.7       1.4       0.383  15.3
#>  3 Dog                        Canis         carni Carnivora       domesticated        10.1       2.9       0.333  13.9
#>  4 Long-nosed armadillo       Dasypus       carni Cingulata       lc                  17.4       3.1       0.383   6.6
#>  5 Domestic cat               Felis         carni Carnivora       domesticated        12.5       3.2       0.417  11.5
#>  6 Pilot whale                Globicephalus carni Cetacea         cd                   2.7       0.1      NA      21.4
#>  7 Gray seal                  Haliochoerus  carni Carnivora       lc                   6.2       1.5      NA      17.8
#>  8 Thick-tailed opposum       Lutreolina    carni Didelphimorphia lc                  19.4       6.6      NA       4.6
#>  9 Slow loris                 Nyctibeus     carni Primates        NA                  11        NA        NA      13  
#> 10 Northern grasshopper mouse Onychomys     carni Rodentia        lc                  14.5      NA        NA       9.5
#> 11 Tiger                      Panthera      carni Carnivora       en                  15.8      NA        NA       8.2
#> 12 Jaguar                     Panthera      carni Carnivora       nt                  10.4      NA        NA      13.6
#> 13 Lion                       Panthera      carni Carnivora       vu                  13.5      NA        NA      10.5
#> 14 Caspian seal               Phoca         carni Carnivora       vu                   3.5       0.4      NA      20.5
#> 15 Common porpoise            Phocoena      carni Cetacea         vu                   5.6      NA        NA      18.4
#> 16 Bottle-nosed dolphin       Tursiops      carni Cetacea         NA                   5.2      NA        NA      18.8
#> 17 Genet                      Genetta       carni Carnivora       NA                   6.3       1.3      NA      17.7
#> 18 Arctic fox                 Vulpes        carni Carnivora       NA                  12.5      NA        NA      11.5
#> 19 Red fox                    Vulpes        carni Carnivora       NA                   9.8       2.4       0.35   14.2
#>    brainwt  bodywt
#>      <dbl>   <dbl>
#>  1 NA       50    
#>  2 NA       20.5  
#>  3  0.07    14    
#>  4  0.0108   3.5  
#>  5  0.0256   3.3  
#>  6 NA      800    
#>  7  0.325   85    
#>  8 NA        0.37 
#>  9  0.0125   1.4  
#> 10 NA        0.028
#> 11 NA      163.   
#> 12  0.157  100    
#> 13 NA      161.   
#> 14 NA       86    
#> 15 NA       53.2  
#> 16 NA      173.   
#> 17  0.0175   2    
#> 18  0.0445   3.38 
#> 19  0.0504   4.23 
#> 
#> $herbi
#> # A tibble: 32 x 11
#>    name             genus       vore  order          conservation sleep_total sleep_rem sleep_cycle awake brainwt
#>    <chr>            <chr>       <chr> <chr>          <chr>              <dbl>     <dbl>       <dbl> <dbl>   <dbl>
#>  1 Mountain beaver  Aplodontia  herbi Rodentia       nt                  14.4       2.4      NA       9.6 NA     
#>  2 Cow              Bos         herbi Artiodactyla   domesticated         4         0.7       0.667  20    0.423 
#>  3 Three-toed sloth Bradypus    herbi Pilosa         NA                  14.4       2.2       0.767   9.6 NA     
#>  4 Roe deer         Capreolus   herbi Artiodactyla   lc                   3        NA        NA      21    0.0982
#>  5 Goat             Capri       herbi Artiodactyla   lc                   5.3       0.6      NA      18.7  0.115 
#>  6 Guinea pig       Cavis       herbi Rodentia       domesticated         9.4       0.8       0.217  14.6  0.0055
#>  7 Chinchilla       Chinchilla  herbi Rodentia       domesticated        12.5       1.5       0.117  11.5  0.0064
#>  8 Tree hyrax       Dendrohyrax herbi Hyracoidea     lc                   5.3       0.5      NA      18.7  0.0123
#>  9 Asian elephant   Elephas     herbi Proboscidea    en                   3.9      NA        NA      20.1  4.60  
#> 10 Horse            Equus       herbi Perissodactyla domesticated         2.9       0.6       1      21.1  0.655 
#>      bodywt
#>       <dbl>
#>  1    1.35 
#>  2  600    
#>  3    3.85 
#>  4   14.8  
#>  5   33.5  
#>  6    0.728
#>  7    0.42 
#>  8    2.95 
#>  9 2547    
#> 10  521    
#> # ... with 22 more rows
#> 
#> $insecti
#> # A tibble: 5 x 11
#>   name                  genus        vore    order        conservation sleep_total sleep_rem sleep_cycle awake brainwt
#>   <chr>                 <chr>        <chr>   <chr>        <chr>              <dbl>     <dbl>       <dbl> <dbl>   <dbl>
#> 1 Big brown bat         Eptesicus    insecti Chiroptera   lc                  19.7       3.9       0.117   4.3 0.0003 
#> 2 Little brown bat      Myotis       insecti Chiroptera   NA                  19.9       2         0.2     4.1 0.00025
#> 3 Giant armadillo       Priodontes   insecti Cingulata    en                  18.1       6.1      NA       5.9 0.081  
#> 4 Eastern american mole Scalopus     insecti Soricomorpha lc                   8.4       2.1       0.167  15.6 0.0012 
#> 5 Short-nosed echidna   Tachyglossus insecti Monotremata  NA                   8.6      NA        NA      15.4 0.025  
#>   bodywt
#>    <dbl>
#> 1  0.023
#> 2  0.01 
#> 3 60    
#> 4  0.075
#> 5  4.5  
#> 
#> $omni
#> # A tibble: 20 x 11
#>    name                       genus         vore  order           conservation sleep_total sleep_rem sleep_cycle awake
#>    <chr>                      <chr>         <chr> <chr>           <chr>              <dbl>     <dbl>       <dbl> <dbl>
#>  1 Owl monkey                 Aotus         omni  Primates        NA                  17         1.8      NA       7  
#>  2 Greater short-tailed shrew Blarina       omni  Soricomorpha    lc                  14.9       2.3       0.133   9.1
#>  3 Grivet                     Cercopithecus omni  Primates        lc                  10         0.7      NA      14  
#>  4 Star-nosed mole            Condylura     omni  Soricomorpha    lc                  10.3       2.2      NA      13.7
#>  5 African giant pouched rat  Cricetomys    omni  Rodentia        NA                   8.3       2        NA      15.7
#>  6 Lesser short-tailed shrew  Cryptotis     omni  Soricomorpha    lc                   9.1       1.4       0.15   14.9
#>  7 North American Opossum     Didelphis     omni  Didelphimorphia lc                  18         4.9       0.333   6  
#>  8 European hedgehog          Erinaceus     omni  Erinaceomorpha  lc                  10.1       3.5       0.283  13.9
#>  9 Patas monkey               Erythrocebus  omni  Primates        lc                  10.9       1.1      NA      13.1
#> 10 Galago                     Galago        omni  Primates        NA                   9.8       1.1       0.55   14.2
#> 11 Human                      Homo          omni  Primates        NA                   8         1.9       1.5    16  
#> 12 Macaque                    Macaca        omni  Primates        NA                  10.1       1.2       0.75   13.9
#> 13 Chimpanzee                 Pan           omni  Primates        NA                   9.7       1.4       1.42   14.3
#> 14 Baboon                     Papio         omni  Primates        NA                   9.4       1         0.667  14.6
#> 15 Potto                      Perodicticus  omni  Primates        lc                  11        NA        NA      13  
#> 16 African striped mouse      Rhabdomys     omni  Rodentia        NA                   8.7      NA        NA      15.3
#> 17 Squirrel monkey            Saimiri       omni  Primates        NA                   9.6       1.4      NA      14.4
#> 18 Pig                        Sus           omni  Artiodactyla    domesticated         9.1       2.4       0.5    14.9
#> 19 Tenrec                     Tenrec        omni  Afrosoricida    NA                  15.6       2.3      NA       8.4
#> 20 Tree shrew                 Tupaia        omni  Scandentia      NA                   8.9       2.6       0.233  15.1
#>     brainwt bodywt
#>       <dbl>  <dbl>
#>  1  0.0155   0.48 
#>  2  0.00029  0.019
#>  3 NA        4.75 
#>  4  0.001    0.06 
#>  5  0.0066   1    
#>  6  0.00014  0.005
#>  7  0.0063   1.7  
#>  8  0.0035   0.77 
#>  9  0.115   10    
#> 10  0.005    0.2  
#> 11  1.32    62    
#> 12  0.179    6.8  
#> 13  0.44    52.2  
#> 14  0.18    25.2  
#> 15 NA        1.1  
#> 16 NA        0.044
#> 17  0.02     0.743
#> 18  0.18    86.2  
#> 19  0.0026   0.9  
#> 20  0.0025   0.104
#> 
# }