Edgar Anderson's Iris Data in long format.

Edgar Anderson's Iris Data in long format.

iris_long

Format

A data frame with 600 rows and 5 variables

id. Dummy identity number for each flower (150 flowers in total).

Species. The species are Iris setosa, versicolor, and virginica.

attribute. What attribute is being measured ("Sepal" or "Pepal").

measure. What aspect of the attribute is being measured ("Length" or "Width").

value. Value of the measurement.

Source

https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/iris.html

Details

This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica.

This is a modified dataset from datasets package.

Examples

dim(iris_long)

#> [1] 600   5

head(iris_long)

#> # A tibble: 6 x 5
#>      id Species attribute measure value
#>   <int> <fct>   <fct>     <fct>   <dbl>
#> 1     1 setosa  Sepal     Length  5.100
#> 2     2 setosa  Sepal     Length  4.9  
#> 3     3 setosa  Sepal     Length  4.7  
#> 4     4 setosa  Sepal     Length  4.600
#> 5     5 setosa  Sepal     Length  5    
#> 6     6 setosa  Sepal     Length  5.4