Movie information and user ratings from IMDB.com (wide format).

Movie information and user ratings from IMDB.com (wide format).

movies_wide

Format

A data frame with 1813 rows and 14 variables

title. Title of the movie.

year. Year of release.

budget. Total budget (if known) in US dollars

length. Length in minutes.

rating. Average IMDB user rating.

votes. Number of IMDB users who rated this movie.

mpaa. MPAA rating.

action, animation, comedy, drama, documentary, romance, short. Binary variables representing if movie was classified as belonging to that genre.

Source

https://CRAN.R-project.org/package=ggplot2movies

Details

Modified dataset from ggplot2movies package.

The internet movie database, http://imdb.com/, is a website devoted to collecting movie data supplied by studios and fans. It claims to be the biggest movie database on the web and is run by amazon. More about information imdb.com can be found online, http://imdb.com/help/show_leaf?about, including information about the data collection process, http://imdb.com/help/show_leaf?infosource.

Movies were selected for inclusion if they had a known length and had been rated by at least one imdb user.

Examples

dim(movies_wide)

#> [1] 1813   14

head(movies_wide)

#> # A tibble: 6 x 14
#>   title  year length budget rating votes mpaa  Action Animation Comedy Drama
#>   <fct> <int>  <int>  <dbl>  <dbl> <int> <fct>  <int>     <int>  <int> <int>
#> 1 'Til~  1997    113   23      4.8   799 PG-13      0         0      1     0
#> 2 10 T~  1999     97   16      6.7 19095 PG-13      0         0      1     0
#> 3 100 ~  2002     98    1.1    5.6   181 R          0         0      1     0
#> 4 13 G~  2004     98   37      6.4  7859 PG-13      0         0      1     1
#> 5 13th~  1999    102   85      6.1 14344 R          1         0      0     0
#> 6 15 M~  2001    120   42      6.1 10866 R          0         0      0     1
#> # ... with 3 more variables: Documentary <int>, Romance <int>, Short <int>