Multiple pairwise comparison tests with tidy data

Calculate parametric, non-parametric, robust, and Bayes Factor pairwise comparisons between group levels with corrections for multiple testing.

Usage

pairwise_comparisons(
  data,
  x,
  y,
  subject.id = NULL,
  type = "parametric",
  paired = FALSE,
  var.equal = FALSE,
  tr = 0.2,
  bf.prior = 0.707,
  p.adjust.method = "holm",
  k = 2L,
  ...
)

Arguments

data

A dataframe (or a tibble) from which variables specified are to be taken. Other data types (e.g., matrix,table, array, etc.) will not be accepted.

x

The grouping (or independent) variable from the dataframe data. In case of a repeated measures or within-subjects design, if subject.id argument is not available or not explicitly specified, the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted, the results can be inaccurate when there are more than two levels in x and there are NAs present. The data is expected to be sorted by user in subject-1,subject-2, ..., pattern.

y

The response (or outcome or dependent) variable from the dataframe data.

subject.id

Relevant in case of a repeated measures or within-subjects design (paired = TRUE, i.e.), it specifies the subject or repeated measures identifier. Important: Note that if this argument is NULL (which is the default), the function assumes that the data has already been sorted by such an id by the user and creates an internal identifier. So if your data is not sorted and you leave this argument unspecified, the results can be inaccurate when there are more than two levels in x and there are NAs present.

type

A character specifying the type of statistical approach:

"parametric"

"nonparametric"

"robust"

"bayes"

You can specify just the initial letter.

paired

Logical that decides whether the experimental design is repeated measures/within-subjects or between-subjects. The default is FALSE.

var.equal

a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

tr

Trim level for the mean when carrying out robust tests. In case of an error, try reducing the value of tr, which is by default set to 0.2. Lowering the value might help.

bf.prior

A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors and posterior estimates. In addition to numeric arguments, several named values are also recognized: "medium", "wide", and "ultrawide", corresponding to r scale values of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value corresponds to scale for fixed effects.

p.adjust.method

Adjustment method for p-values for multiple comparisons. Possible methods are: "holm" (default), "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".

k

Number of digits after decimal point (should be an integer) (Default: k = 2L).

...

Additional arguments passed to other methods.

Value

A tibble dataframe containing two columns corresponding to group levels being compared with each other (group1 and group2) and p.valuecolumn corresponding to this comparison. The dataframe will also contain a p.value.label column containing a label for this p-value, in case this needs to be displayed in ggsignif::geom_ggsignif. In addition to these common columns across the different types of statistics, there will be additional columns specific to the type of test being run. This function provides a unified syntax to carry out pairwise comparison tests and internally relies on other packages to carry out these tests. For more details about the included tests, see the documentation for the respective functions:

parametric : stats::pairwise.t.test() (paired) and PMCMRplus::gamesHowellTest() (unpaired)

non-parametric : PMCMRplus::durbinAllPairsTest() (paired) and PMCMRplus::kwAllPairsDunnTest() (unpaired)

robust : WRS2::rmmcp() (paired) and WRS2::lincon() (unpaired)

Bayes Factor : BayesFactor::ttestBF()

References

For more, see: https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/pairwise.html

Examples

# \donttest{
if (require("PMCMRplus")) {
  # for reproducibility
  set.seed(123)
  library(ggstatsplot)
  library(statsExpressions) # for data

  # show all columns and make the column titles bold
  # as a user, you don't need to do this; this is just for the package website
  options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE, pillar.subtle_num = TRUE)

  #------------------- between-subjects design ----------------------------

  # parametric
  # if `var.equal = TRUE`, then Student's t-test will be run
  pairwise_comparisons(
    data            = mtcars,
    x               = cyl,
    y               = wt,
    type            = "parametric",
    var.equal       = TRUE,
    paired          = FALSE,
    p.adjust.method = "none"
  )

  # if `var.equal = FALSE`, then Games-Howell test will be run
  pairwise_comparisons(
    data            = mtcars,
    x               = cyl,
    y               = wt,
    type            = "parametric",
    var.equal       = FALSE,
    paired          = FALSE,
    p.adjust.method = "bonferroni"
  )

  # non-parametric (Dunn test)
  pairwise_comparisons(
    data            = mtcars,
    x               = cyl,
    y               = wt,
    type            = "nonparametric",
    paired          = FALSE,
    p.adjust.method = "none"
  )

  # robust (Yuen's trimmed means *t*-test)
  pairwise_comparisons(
    data            = mtcars,
    x               = cyl,
    y               = wt,
    type            = "robust",
    paired          = FALSE,
    p.adjust.method = "fdr"
  )

  # Bayes Factor (Student's *t*-test)
  pairwise_comparisons(
    data   = mtcars,
    x      = cyl,
    y      = wt,
    type   = "bayes",
    paired = FALSE
  )

  #------------------- within-subjects design ----------------------------

  # parametric (Student's *t*-test)
  pairwise_comparisons(
    data            = bugs_long,
    x               = condition,
    y               = desire,
    subject.id      = subject,
    type            = "parametric",
    paired          = TRUE,
    p.adjust.method = "BH"
  )

  # non-parametric (Durbin-Conover test)
  pairwise_comparisons(
    data            = bugs_long,
    x               = condition,
    y               = desire,
    subject.id      = subject,
    type            = "nonparametric",
    paired          = TRUE,
    p.adjust.method = "BY"
  )

  # robust (Yuen's trimmed means t-test)
  pairwise_comparisons(
    data            = bugs_long,
    x               = condition,
    y               = desire,
    subject.id      = subject,
    type            = "robust",
    paired          = TRUE,
    p.adjust.method = "hommel"
  )

  # Bayes Factor (Student's *t*-test)
  pairwise_comparisons(
    data       = bugs_long,
    x          = condition,
    y          = desire,
    subject.id = subject,
    type       = "bayes",
    paired     = TRUE
  )
}
#> # A tibble: 6 × 20
#>   group1 group2 term       effectsize      estimate conf.level conf.low
#>   <chr>  <chr>  <chr>      <chr>              <dbl>      <dbl>    <dbl>
#> 1 HDHF   HDLF   Difference Bayesian t-test    1.10        0.95   0.503 
#> 2 HDHF   LDHF   Difference Bayesian t-test    0.455       0.95  -0.0659
#> 3 HDHF   LDLF   Difference Bayesian t-test    2.13        0.95   1.63  
#> 4 HDLF   LDHF   Difference Bayesian t-test   -0.661       0.95  -1.31  
#> 5 HDLF   LDLF   Difference Bayesian t-test    0.991       0.95   0.385 
#> 6 LDHF   LDLF   Difference Bayesian t-test    1.65        0.95   1.16  
#>   conf.high    pd rope.percentage prior.distribution prior.location prior.scale
#>       <dbl> <dbl>           <dbl> <chr>                       <dbl>       <dbl>
#> 1    1.77   1.00            0     cauchy                          0       0.707
#> 2    0.931  0.962           0.194 cauchy                          0       0.707
#> 3    2.63   1               0     cauchy                          0       0.707
#> 4    0.0351 0.97            0.147 cauchy                          0       0.707
#> 5    1.58   0.999           0     cauchy                          0       0.707
#> 6    2.17   1               0     cauchy                          0       0.707
#>       bf10 method          log_e_bf10 n.obs expression  
#>      <dbl> <chr>                <dbl> <int> <list>      
#> 1 4.16e+ 1 Bayesian t-test      3.73     88 <expression>
#> 2 5.83e- 1 Bayesian t-test     -0.539    88 <expression>
#> 3 1.20e+10 Bayesian t-test     23.2      88 <expression>
#> 4 6.98e- 1 Bayesian t-test     -0.359    88 <expression>
#> 5 1.81e+ 1 Bayesian t-test      2.90     88 <expression>
#> 6 4.81e+ 6 Bayesian t-test     15.4      88 <expression>
#>   label                           test.details    
#>   <chr>                           <chr>           
#> 1 list(~log[e](BF['01'])==-3.73)  Student's t-test
#> 2 list(~log[e](BF['01'])==0.54)   Student's t-test
#> 3 list(~log[e](BF['01'])==-23.21) Student's t-test
#> 4 list(~log[e](BF['01'])==0.36)   Student's t-test
#> 5 list(~log[e](BF['01'])==-2.90)  Student's t-test
#> 6 list(~log[e](BF['01'])==-15.39) Student's t-test
# }