Making text subtitle for contingency analysis (Pearson's chi-square test for independence for between-subjects design or McNemar's test for within-subjects design)

Making text subtitle for contingency analysis (Pearson's chi-square test for independence for between-subjects design or McNemar's test for within-subjects design)

subtitle_contigency_tab(data, main, condition, counts = NULL,
  nboot = 25, paired = FALSE, stat.title = NULL, conf.level = 0.95,
  conf.type = "norm", messages = TRUE, k = 3)

Arguments

data
The data as a data frame (matrix or tables will not be accepted).

main
The variable to use as the rows in the contingency table.

condition
The variable to use as the columns in the contingency table.

counts
A string naming a variable in data containing counts, or NULL if each row represents a single observation (Default).

nboot
Number of bootstrap samples for computing effect size (Default: 25).

paired
Logical indicating whether data came from a within-subjects design study (Default: FALSE). If TRUE, McNemar test subtitle will be returned. If FALSE, Pearson's chi-square test will be returned.

stat.title
Title for the effect being investigated with the chi-square test. The default is NULL, i.e. no title will be added to describe the effect being shown. An example of a stat.title argument will be something like "main x condition" or "interaction".

conf.level
Scalar between 0 and 1. If NULL, the defaults return 95 lower and upper confidence intervals (0.95).

conf.type
A vector of character strings representing the type of intervals required. The value should be any subset of the values "norm", "basic", "perc", "bca". For more, see ?boot::boot.ci.

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

k
Number of digits after decimal point (should be an integer) (Default: k = 3).

Examples

# without counts data
subtitle_contigency_tab(
  data = mtcars,
  main = am,
  condition = cyl
)

#> [32mNote:[39m [34m95% CI for Cramer's V was computed with  [33m25[34m bootstrap samples.[39m

#> 

#> paste(NULL, italic(chi)^2, "(", 2L, ") = ", "8.741", ", ", italic("p"), 
#>     " = ", "0.013", ", ", italic(V), " = ", "0.523", ", 95% CI [", 
#>     "0.214", ", ", "0.783", "]", ", ", italic("n"), " = ", 32)

# with counts data
# in case of no variation, NaN will be shown for results
library(jmv)

dat <- as.data.frame(HairEyeColor) %>%
  dplyr::filter(.data = ., Sex == "Male")

subtitle_contigency_tab(
  data = dat,
  main = Hair,
  condition = Sex,
  counts = Freq
)

#> paste(NULL, italic(chi)^2, "(", 3L, ") = ", "NaN", ", ", italic("p"), 
#>     " = ", "NaN", ", ", italic(V), " = ", "NaN", ", 95% CI [", 
#>     "NaN", ", ", "NaN", "]", ", ", italic("n"), " = ", 279)