Making text subtitle for the correlation test.

Making text subtitle for the correlation test.

subtitle_ggscatterstats(data, x, y, nboot = 100, beta = 0.1,
  type = "pearson", bf.prior = 0.707, conf.level = 0.95,
  conf.type = "norm", messages = TRUE, k = 3)

Arguments

data
Dataframe from which variables specified are preferentially to be taken.

x
A vector containing the explanatory variable.

y
The response - a vector of length the number of rows of x.

nboot
Number of bootstrap samples for computing effect size (Default: 100).

beta
bending constant (Default: 0.1). For more, see ?WRS2::pbcor.

type
Type of association between paired samples required (""parametric": Pearson's product moment correlation coefficient" or ""nonparametric": Spearman's rho" or ""robust": percentage bend correlation coefficient" or ""bayes": Bayes Factor for Pearson's r"). Corresponding abbreviations are also accepted: "p" (for parametric/pearson's), "np" (nonparametric/spearman), "r" (robust), "bf" (for bayes factor), resp.

bf.prior
A number between 0.5 and 2 (default 0.707), the prior width to use in calculating Bayes factors.

conf.level
Scalar between 0 and 1. If NULL, the defaults return 95 lower and upper confidence intervals (0.95).

conf.type
A vector of character strings representing the type of intervals required. The value should be any subset of the values "norm", "basic", "perc", "bca". For more, see ?boot::boot.ci.

messages
Decides whether messages references, notes, and warnings are to be displayed (Default: TRUE).

k
Number of digits after decimal point (should be an integer) (Default: k = 3).

Examples

# without changing defaults
subtitle_ggscatterstats(
  data = ggplot2::midwest,
  x = area,
  y = percblack
)

#> paste("Pearson's ", italic("r"), "(", 435L, ")", " = ", "-0.046", 
#>     ", 95% CI [", "-0.140", ", ", "0.048", "], ", italic("p"), 
#>     " = ", "0.332", ", ", italic("n"), " = ", 437L)

# changing defaults
subtitle_ggscatterstats(
  data = ggplot2::midwest,
  x = area,
  y = percblack,
  nboot = 25,
  beta = 0.2,
  type = "r",
  k = 1
)

#> Note: 95% CI for percentage bend correlation was computed with 25 bootstrap samples.

#> 

#> paste("robust ", italic(r), " = ", "0.0", ", 95% CI [", "-0.1", 
#>     ", ", "0.1", "], ", italic("p"), " = ", "0.960", ", ", italic("n"), 
#>     " = ", 437L)