Making text subtitle for Proportion Test (N Outcomes)

This is going to be a chi-squared Goodness of fit test.

subtitle_onesample_proptest(data, main, counts = NULL, ratio = NULL,
  conf.level = 0.95, stat.title = NULL, legend.title = NULL, k = 2,
  ...)

Arguments

data
A dataframe (or a tibble) from which variables specified are to be taken. A matrix or tables will not be accepted.

main
The variable to use as the rows in the contingency table.

counts
A string naming a variable in data containing counts, or NULL if each row represents a single observation (Default).

ratio
A vector of numbers: the expected proportions for the proportion test. Default is NULL, which means if there are two levels ratio = c(1,1), etc.

conf.level
Scalar between 0 and 1. If unspecified, the defaults return 95% lower and upper confidence intervals (0.95).

stat.title
Title for the effect being investigated with the chi-square test. The default is NULL, i.e. no title will be added to describe the effect being shown. An example of a stat.title argument will be something like "main x condition" or "interaction".

legend.title
Title text for the legend.

k
Number of digits after decimal point (should be an integer) (Default: k = 2).

...
Additional arguments (currently ignored).

Examples

# for reproducibility
set.seed(123)
library(jmv)

# with counts
subtitle_onesample_proptest(
  data = as.data.frame(HairEyeColor),
  main = Eye,
  counts = Freq
)

#> paste(NULL, italic(chi)^2, "(", "3", ") = ", "133.47", ", ", 
#>     italic("p"), " = ", "< 0.001", ", ", italic("V")["Cramer"], 
#>     " = ", "0.27", ", CI"["95%"], " [", "0.23", ", ", "0.32", 
#>     "]", ", ", italic("n"), " = ", 592L)

# in case of no variation, only sample size will be shown
subtitle_onesample_proptest(
  data = cbind.data.frame(x = rep("a", 10)),
  main = x
)

#> Error in chisq.test(counts, p = expProps) : 
#>   'x' must at least have 2 elements
#> Warning: Proportion test will not be run because it requires x to have at least 
#> 2 levels with non-zero frequencies.

#> 

#> paste(italic("n"), " = ", 10L)